% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adae_risk_summary.R
\name{adae_risk_summary}
\alias{adae_risk_summary}
\title{ADAE Summary with Risk Statistics}
\usage{
adae_risk_summary(
  datain,
  a_subset = NA_character_,
  summary_by = "Patients",
  hterm = "AEBODSYS",
  lterm = "AEDECOD",
  ctrlgrp,
  trtgrp,
  statistics = "Risk Ratio",
  alpha = 0.05,
  cutoff = 5,
  sort_opt = "Ascending",
  sort_var = "Count"
)
}
\arguments{
\item{datain}{Input dataset after pre_processing and running \code{mentry()} to \emph{ADAE} data}

\item{a_subset}{Analysis Subset condition specific to categorical analysis.}

\item{summary_by}{Measure to construct the summary by. Values: \verb{'Patients' or 'Events'}.}

\item{hterm}{High Level Adverse Event term variable, used for analysis}

\item{lterm}{Low Level Adverse Event term variable, used for analysis}

\item{ctrlgrp}{Treatment Control value.}

\item{trtgrp}{Treatment(s) to create pairs. Only 1 value for Volcano/table, can be multiple
for \code{forest_plot()}.}

\item{statistics}{Statistic to be calculated. Values: \verb{'Risk Ratio' or 'Risk Difference'}.}

\item{alpha}{Alpha value to determine confidence interval for risk calculation. Default: \code{0.05}}

\item{cutoff}{Incidence Cutoff Value; consider only terms with \verb{incidence percentage > cutoff}.}

\item{sort_opt}{How to sort terms, only for table/forest plot.
Values: \verb{'Ascending','Descending','Alphabetical'}.}

\item{sort_var}{Metric to sort by. Values: \verb{'Count','Percent','RiskValue'}.}
}
\value{
List of summarized data frames for Adverse Events based on high and lower term.
}
\description{
ADAE Summary with Risk Statistics
}
\examples{
data(adae)

ae_pre_process <- ae_pre_processor(
  datain = adae,
  ae_filter = "Any Event",
  obs_residual = 0,
  fmq_data = FMQ_Consolidated_List
)
ae_entry <- ae_pre_process[["data"]] |>
  mentry(
    subset = NA,
    byvar = "AEBODSYS",
    trtvar = "TRTA",
    trtsort = "TRTAN",
    trttotalyn = "N",
    add_grpmiss = "N",
    sgtotalyn = "N",
    pop_fil = "Overall Population"
  )

ae_risk <- ae_entry |>
  adae_risk_summary(
    a_subset = ae_pre_process[["a_subset"]],
    summary_by = "Patients",
    hterm = "AEBODSYS",
    lterm = "AEDECOD",
    ctrlgrp = "Placebo",
    trtgrp = "Xanomeline Low Dose",
    statistics = "Risk Ratio",
    alpha = 0.05,
    cutoff = 5,
    sort_opt = "Ascending",
    sort_var = "Count"
  )

ae_risk |>
  tbl_processor(keepvars = c("Risk Ratio (CI)", "P-value")) |>
  tbl_display()

}
