% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_merge.R
\name{adsl_merge}
\alias{adsl_merge}
\title{Merge adsl dataset with the analysis dataset}
\usage{
adsl_merge(adsl = NULL, adsl_subset = "", dataset_add = NULL, byvars = NULL)
}
\arguments{
\item{adsl}{adsl dataset}

\item{adsl_subset}{population variable subset condition}

\item{dataset_add}{analysis dataset}

\item{byvars}{Variables to merge the datasets by}
}
\value{
merged dataset
}
\description{
Merge adsl dataset with the analysis dataset
}
\examples{
data(lab_data)

adsl_merge(
  adsl = lab_data$adsl,
  adsl_subset = "SAFFL=='Y'",
  dataset_add = lab_data$adlb
)

}
