% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsl_summary.R
\name{adsl_summary}
\alias{adsl_summary}
\title{Demographic Characteristics Table}
\usage{
adsl_summary(
  datain,
  vars,
  stat_vars = "n~minmaxc~mean(sd)~median~q1q3",
  pctdisp = "TRT",
  total_catyn = "N",
  total_catlabel = "Total",
  miss_catyn = "N",
  miss_catlabel = "Missing",
  pctsyn = "Y",
  sigdec_stat = 2,
  sigdec_cat = 2,
  a_subset = NA_character_,
  denom_subset = NA_character_,
  sparseyn = "N",
  sparsebyvalyn = "N"
)
}
\arguments{
\item{datain}{Input data from \code{mentry()} output}

\item{vars}{Names of \code{adsl} variables to display (Add \code{"-S"} to numeric variables),
tilde-separated}

\item{stat_vars}{Statistics to display in table for numeric vars, tilde-separated.}

\item{pctdisp}{Method to calculate denominator (for \%) by.
Possible values: \code{"TRT"}, \code{"VAR"}, \code{"COL"}, \code{"SUBGRP"}, \code{"CAT"}, \code{"NONE"}, \code{"NO"}, \code{"DPTVAR"},
\code{"BYVARxyN"}}

\item{total_catyn}{To return a 'Total' row for the categories of \code{dptvar}
variable or not. Possible values: \code{"Y"/"N"}}

\item{total_catlabel}{Label for total category row. eg- "All"/"Total"}

\item{miss_catyn}{To include empty/blank values as \code{miss_catlabel} in categories of
\code{dptvar} variable or not. Values: \code{"Y"/"N"}}

\item{miss_catlabel}{Label for missing values}

\item{pctsyn}{Display Percentage Sign in table or not. Values: \code{"Y"/"N"}}

\item{sigdec_stat}{Number of base decimal places to retain in output of summary statistic.
Applies to mean, min, max, sd etc}

\item{sigdec_cat}{Number of decimal places for \% displayed in output}

\item{a_subset}{Analysis Subset condition; tilde-separated for each variable in \code{vars}.}

\item{denom_subset}{Subset condition to be applied to dataset for calculating denominator,
tilde-separated for categorical variables within \code{vars}.}

\item{sparseyn}{To sparse missing categories/treatments or not? \code{"Y"/"N"}}

\item{sparsebyvalyn}{Sparse missing categories within by groups. \code{"Y"/"N"}}
}
\value{
\code{data.frame} to be passed on to \code{tbl_processor} and \code{tbl_display}
}
\description{
Demographic Characteristics Table
}
\details{
\itemize{
\item \code{vars} should contain all variables for analysis with \code{"-S"} added to numeric/summary stat
variables, the remaining being for categorical analysis.
eg. for \code{"AGEGR1/AGEGR1N~AGE-S~SEX/SEXN~BMIBL-S"}, \code{AGEGR1} and \code{SEX} will be analysed by
category and \code{AGE} and \code{BMIBL} as summary statistics.
\item Argument \code{stat_vars} should contain names of statistic to apply to all summary analysis
variables. \code{sigdec} applies only to statistical analysis of numeric variables (-S)
\item Arguments \code{pctdisp}, \code{total_catyn}, \code{miss_catyn}, \code{miss_catlabel} apply to all variables
under categorical analyses.
\item \code{a_subset} should tilde-separated subset conditions, corresponding to each variable in
\code{vars}. Length should be \code{1} or number of variables in \code{vars}.
If only \code{1} condition is given, it will apply to all variables.
\item \code{denom_subset}, commonly used only when \code{pctdisp} is \code{"SUBGRP"} or \code{"DPTVAR"} corresponds
to categorical variables within \code{vars}. eg. for \code{"AGEGR1~AGE-S~SEX~BMIBL-S"}, there should be \code{2}
tilde-separated conditions like \code{"!is.na(USUBJID)~!is.na(SITEID)"}.
If only \code{1} condition is given, it will apply to all categorical/discrete variables.
}
}
\examples{
data(adsl)

mentry_df <- adsl |>
  mentry(
    subset = NA_character_,
    byvar = NA_character_,
    trtvar = "TRT01A",
    trtsort = "TRT01AN",
    subgrpvar = NA_character_,
    trttotalyn = "Y",
    add_grpmiss = "N",
    pop_fil = "SAFFL"
  )

adsl_sum <- mentry_df |>
  adsl_summary(
    vars = "AGEGR1/AGEGR1N~AGE-S~SEX/SEXN~RACE/RACEN",
    a_subset = "AGE<65~AGE>80~SEX=='F'~NA"
  )

adsl_sum |>
  display_bign_head(mentry_data = mentry_df) |>
  tbl_processor(
    statlabel = "N~Range~Meansd~Median~Q1Q3",
    dptlabel = "Age Group~_NONE_~Sex~Race",
    addrowvar = "DPTVAR"
  ) |>
  tbl_display() |>
  flextable::autofit()

# Same variable with 2 unique subset conditions
adsl_sum <- mentry_df |>
  adsl_summary(
    vars = "AGEGR1~AGE-S~SEX~SEX~RACE",
    a_subset = "AGE<65~AGE>80~SEX=='F'~SEX=='M'~NA"
  )

adsl_sum |>
  display_bign_head(mentry_data = mentry_df) |>
  tbl_processor(
    statlabel = "N~Range~Meansd~Median~IQR",
    dptlabel = "Age Group~_NONE_~Sex1~Sex2~Race",
    addrowvar = "DPTVAR"
  ) |>
  tbl_display() |>
  flextable::autofit()

# participants in each data analysis set
mentry_df |>
  adsl_summary(
    vars = "ITTFL~SAFFL~DISCONFL",
    a_subset = "ITTFL=='Y'~SAFFL=='Y'~DISCONFL=='Y'"
  ) |>
  display_bign_head(mentry_df) |>
  tbl_processor(
    statlabel = "N",
    dptlabel = "Intent-to-treat (n\%)~Safety (n\%)~Discontinued (n\%)",
    disp_value_col = "N"
  ) |>
  tbl_display() |>
  flextable::autofit()

# with groups and subgroups
mentry_df_grouped <- adsl |>
  mentry(
    subset = NA_character_,
    byvar = "SEX",
    trtvar = "TRT01A",
    trtsort = "TRT01AN",
    subgrpvar = "ETHNIC",
    trttotalyn = "N",
    add_grpmiss = "N",
    sgtotalyn = "Y",
    pop_fil = "SAFFL"
  )

adsl_sum_grouped <- mentry_df_grouped |>
  adsl_summary(
    vars = "AGEGR1~AGE-S~RACE"
  )

adsl_sum_grouped |>
  display_bign_head(mentry_df_grouped) |>
  tbl_processor(
    statlabel = "N~Range~Meansd~Median~IQR",
    dptlabel = "Age Group~Age~Race",
    addrowvar = "DPTVAR"
  ) |>
  tbl_display(
    bylabel = "Sex",
  ) |>
  flextable::autofit()

## Demographics Characteristics - Full Analysis Set with `pharmaverseadam` datasets
\dontrun{
library(dplyr)
library(tidyr)

advs <- pharmaverseadam::advs |>
  filter(
    .data$PARAMCD \%in\% c("HEIGHT", "WEIGHT", "BMI"),
    .data$ABLFL == "Y"
  ) |>
  select(all_of(c("USUBJID", "PARAMCD", "AVAL"))) |>
  pivot_wider(
    id_cols = "USUBJID",
    names_from = "PARAMCD",
    values_from = "AVAL"
  )

adsl_vs <- pharmaverseadam::adsl |>
  adsl_merge(
    dataset_add = advs,
    adsl_subset = "toupper(TRT01A) != 'SCREEN FAILURE'"
  ) |>
  mentry(
    trtvar = "TRT01A",
    trttotalyn = "Y"
  )

adsl_vs |>
  adsl_summary(
    vars = "SEX~AGE-S~AGEGR1~RACE~ETHNIC~HEIGHT-S~WEIGHT-S~BMI-S",
    stat_vars = "median(minmax)~mean(sd)"
  ) |>
  display_bign_head(adsl_vs) |>
  tbl_processor(
    dptlabel = "Sex, n(\%)~Age (Years)~Age Category (Years), n(\%)~Race, n(\%)~Ethnicity,
    n(\%)~Height (cm)~Weight (kg)~BMI (kg/m2)",
    statlabel = "Median (Min, Max)~Mean (SD)",
    addrowvars = "DPTVAR"
  ) |>
  tbl_display() |>
  flextable::autofit()
}

}
