% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ae_pre_process}
\alias{ae_pre_process}
\title{ae_pre}
\format{
List of length \code{2}, 1 data frame and 1 string
}
\usage{
ae_pre_process
}
\description{
Pre Processed \code{adae} data from \code{ae_pre_processor()}
}
\keyword{datasets}
