% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_volcano_plot.R
\name{ae_volcano_opts}
\alias{ae_volcano_opts}
\title{Volcano Plot axis Options}
\usage{
ae_volcano_opts(
  datain,
  trt1_label = "Control",
  trt2_label = "Exposure",
  statistic = "Risk Ratio",
  pvalue_trans = "none",
  xref_offset = 1
)
}
\arguments{
\item{datain}{Input data for volcano plot}

\item{trt1_label}{Control Treatment label}

\item{trt2_label}{Other treatment label}

\item{statistic}{Statistic used for risk analysis}

\item{pvalue_trans}{Required pvalue transformation. Values: "none"/"-log10"}

\item{xref_offset}{Offset value for X axis reference line; offset from 0 is \code{statistic} is Risk
Difference and from 1 if \code{statistic} is Risk Ratio.}
}
\value{
List of volcano-specific options
}
\description{
Volcano Plot axis Options
}
\examples{
ae_volcano_opts(ae_risk,
  pvalue_trans = "-log10"
)
}
