% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_volcano_plot.R
\name{ae_volcano_plot}
\alias{ae_volcano_plot}
\title{Generate Volcano Plot of Adverse Events from Risk values of treatment pairs}
\usage{
ae_volcano_plot(
  datain,
  axis_opts,
  legend_opts = list(label = "", pos = "bottom", dir = "horizontal"),
  xref = c(1, 0, 2),
  pvalue_sig = 0.05,
  interactive = "N"
)
}
\arguments{
\item{datain}{(\code{data.frame})\cr Input dataset containing statistics for treatment
pairs from AE data, generated by \code{risk_stat()}.}

\item{axis_opts}{A \code{list} of axis specific options retrieved from \code{plot_axis_opts()} and
\code{ae_volcano_opts()}.}

\item{legend_opts}{Legend styling option, a \code{list} containing \code{label}, \code{pos}(position) and
\code{dir} (direction).}

\item{xref}{(\code{numeric})\cr Numeric vector, from \code{ae_volcano_opts} First value will be the main
reference, all drawn vertically.}

\item{pvalue_sig}{(\code{numeric})\cr p-value cut-off, filter values with adjusted \code{p} less than
this argument}

\item{interactive}{Return interactive plot or static. Values: "Y"/"N".}
}
\value{
a ggplot object (volcano plot)
}
\description{
Generate Volcano Plot of Adverse Events from Risk values of treatment pairs
}
\details{
It is recommended not to change \code{pos} and \code{dir} in \code{legend_opts}; additionally, these options
are fixed for the case of \code{interactive} = Y and cannot be changed.
\code{xref} must be of length 2 or more.
Input \code{datain} from \code{risk_stat()} should only contain one treatment pair and not more.
}
\examples{
data("adae")

ae_pre <- ae_pre_processor(
  datain = adae,
  obs_residual = 0,
  fmq_data = NA
)

ae_entry <- mentry(
  datain = ae_pre$data,
  subset = NA,
  byvar = "AEBODSYS",
  trtvar = "TRTA",
  trtsort = "TRTAN",
  subgrpvar = NA,
  trttotalyn = "N",
  add_grpmiss = "N",
  sgtotalyn = "N",
  pop_fil = "SAFFL"
)

ae_risk <- risk_stat(
  datain = ae_entry,
  a_subset = ae_pre$a_subset,
  summary_by = "Patients",
  eventvar = "AEDECOD",
  ctrlgrp = "Placebo",
  trtgrp = "Xanomeline High Dose",
  statistics = "Risk Ratio",
  alpha = 0.05,
  cutoff_where = "FREQ >5",
  sort_opt = "Ascending",
  sort_var = "Count"
)

vaxis_opts <- ae_volcano_opts(
  ae_risk,
  "Control",
  "Exposure",
  pvalue_trans = "-log10"
)

axis_opts <- plot_axis_opts(
  ylinearopts = vaxis_opts$ylinearopts,
  yaxis_scale = vaxis_opts$yaxis_scale,
  xaxis_label = vaxis_opts$xaxis_label,
  yaxis_label = vaxis_opts$yaxis_label
)

ae_risk |>
  ae_volcano_plot(
    axis_opts = axis_opts,
    xref = vaxis_opts$xref
  )

}
