% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptly_utils.R
\name{as_plotly}
\alias{as_plotly}
\title{Standard convert ggplot to plotly object}
\usage{
as_plotly(
  plot,
  height = NULL,
  width = NULL,
  legend_opts = list(label = "", pos = "bottom", dir = "h"),
  axis_opts = list(xopts = list(showticklabels = TRUE, showline = TRUE, mirror = TRUE),
    yopts = list(showticklabels = TRUE, showline = TRUE, mirror = TRUE)),
  hover = "text"
)
}
\arguments{
\item{plot}{Input ggplot object}

\item{height}{Height of interactive plot}

\item{width}{Width of interactive plot}

\item{legend_opts}{Legend styling option, a \code{list} containing \code{label}, \code{pos}(position) and
\code{dir} (direction).}

\item{axis_opts}{Axis styling option. Default is a named list
\code{list(xopts = list(showticklabels = TRUE, showline = TRUE, mirror = TRUE), yopts = list(showticklabels = TRUE, showline = TRUE, mirror = TRUE))} # nolint}

\item{hover}{Tooltip aesthetic}
}
\value{
plotly widget
}
\description{
Standard convert ggplot to plotly object
}
\examples{
library(ggplot2)

gplot <- adsl[1:100, ] |>
  ggplot(aes(x = .data[["RACE"]], y = .data[["AGE"]], text = .data[["USUBJID"]])) +
  geom_point()
as_plotly(
  plot = gplot,
  height = 400,
  width = 400
)
}
