% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_plot.R
\name{box_plot}
\alias{box_plot}
\title{Create box plot - static and interactive output}
\usage{
box_plot(
  datain,
  series_var = "TRTVAR",
  series_labelvar = series_var,
  series_opts,
  axis_opts,
  legend_opts = list(label = "", pos = "bottom", dir = "horizontal"),
  ystat = "mean",
  boxfill = "N",
  box_opts = c(0.7, 0.9),
  griddisplay = "N",
  plot_title = NULL
)
}
\arguments{
\item{datain}{Input dataset from \code{msumstat()} output element \verb{$gsum}}

\item{ystat}{Additional statistic to be plotted as markers. Default: \emph{mean}}

\item{boxfill}{Whether box has fill color or not. Uses \code{series_opts$color} as
box fill colors if "Y" or as box outline colors if "N". ("Y"/"N")}

\item{box_opts}{Vector containing:
\enumerate{
\item Width of individual boxes in plot and
\item Width of the interval between box-groups of X axis. eg. c(0.9, 0.9)
}}
}
\value{
a ggplot object of boxplot with given parameters
}
\description{
Create box plot - static and interactive output
}
\details{
This function utilises data from processing steps to create boxplot.
\enumerate{
\item Statistics required: In \code{msumstat()} function prior, statvar should be
c("median", "q25", "q75", "sd", "whiskerlow", "whiskerup", ystat) or "box" with
ystat representing additional statistic to be calculated/plotted as points.
\item \code{series_labvar} can be, for eg "TRTTXT" if that variable contains
modified labels for series_var (if "TRTVAR") to be shown in legend only
\item \code{color} , \code{shape}, \code{size} elements of \code{series_opts} should be of the same
length as or not less than number of levels/unique values in the variable
\code{series_var} - recommended to get from \code{plot_aes_opts()}
}
}
\examples{
data(adsl)

adsl_entry <- mentry(
  datain = adsl,
  subset = "EFFFL=='Y'",
  byvar = "RACE",
  trtvar = "TRT01A",
  trtsort = "TRT01AN",
  pop_fil = NA
)

adsl_sum <- msumstat(
  datain = adsl_entry,
  dptvar = "AGE",
  statvar = c(
    "mean", "median", "q25", "q75", "whiskerlow",
    "whiskerup"
  )
)[["gsum"]] |>
  plot_display_bign(adsl_entry) |>
  dplyr::mutate(XVAR = BYVAR1)

series_opts <- adsl_sum |>
  plot_aes_opts(
    series_color = c("red", "gold", "cyan"),
    series_shape = c(16, 17, 15),
    series_size = c(1, 1, 1)
  )

box_plot(
  datain = adsl_sum,
  axis_opts = plot_axis_opts(),
  legend_opts = list(
    label = "Treatment", pos = "bottom",
    dir = "horizontal"
  ),
  series_opts = series_opts,
  plot_title = "Age distribution over Race groups"
)
}
