% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carver-package.R
\docType{package}
\name{carver-package}
\alias{carver}
\alias{carver-package}
\title{carver: Clinical Analysis Reporting and Visualization Ensemble in R}
\description{
Development of an open-source tool and package to enable generation of identified interactive plots for clinical review and direct inclusion in submission for regulatory agencies. The scope includes Shiny application for complete report generation from Analysis Model (ADaM) datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pfizer-opensource/carver/}
  \item Report bugs at \url{https://github.com/pfizer-opensource/carver/issues/}
}

}
\author{
\strong{Maintainer}: Smriti Anand \email{smriti.anand@pfizer.com}

Authors:
\itemize{
  \item Sukalpo Saha
  \item Aravind Reddy
}

}
