% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_merge.R
\name{dataset_merge}
\alias{dataset_merge}
\title{Merge Datasets}
\usage{
dataset_merge(..., byvars, subset = NULL, type = "left")
}
\arguments{
\item{...}{Datasets to be merged.}

\item{byvars}{By variables required to perform merge.}

\item{subset}{Dataset specific subset conditions as \code{list}, default is \code{NULL}.
Has to be specified in the same order of datasets to be merged}

\item{type}{Type of join to perform. Values: "left", "right", "inner", "full", "semi", "anti"}
}
\value{
A \code{data.frame}
}
\description{
Merge Datasets
}
\examples{
dataset_merge(
  adsl,
  adlb,
  byvars = "STUDYID~USUBJID~SUBJID",
  subset = list("SEX=='F'", "PARAMCD == 'ALT'")
)

dataset_merge(
  adsl,
  adlb,
  byvars = "STUDYID~USUBJID~SUBJID",
  subset = list("SEX=='F'", NA_character_)
)

dataset_merge(
  adsl,
  adlb,
  byvars = "STUDYID~USUBJID~SUBJID",
  subset = list(NA_character_, "PARAMCD == 'ALT'")
)

dataset_merge(
  adsl,
  adlb,
  byvars = "STUDYID~USUBJID~SUBJID",
  subset = list("USUBJID == '01-701-1015'", NA_character_)
)

## more than 2 datasets

dataset_merge(
  dplyr::filter(adsl, USUBJID == "01-701-1015"),
  adsl,
  adlb,
  byvars = "STUDYID~USUBJID~SUBJID"
)

}
