% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dataset_vignette}
\alias{dataset_vignette}
\title{Output a Dataset in a Vignette}
\usage{
dataset_vignette(df = NULL, disp_vars = NULL, subset = NA_character_)
}
\arguments{
\item{df}{Dataset to output in the vignette}

\item{disp_vars}{Variables selected to demonstrate the outcome of the derivation}

\item{subset}{Subset condition}
}
\value{
A HTML table
}
\description{
Output a Dataset in a Vignette
}
