% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{display_bign_head}
\alias{display_bign_head}
\title{Attach bign value to table columns, modify column header labels}
\usage{
display_bign_head(
  datain,
  mentry_data = NULL,
  trtbignyn = "Y",
  subbignyn = "Y",
  colformat = "",
  notrthead = "Total"
)
}
\arguments{
\item{datain}{Input dataframe for display}

\item{mentry_data}{Input data from \code{mentry()}}

\item{trtbignyn}{Append (N = trt count) to Treatment Column headers? Values: Y/N}

\item{subbignyn}{Append (N = trt*subgrp count) to Subgroup Column headers? Values: Y/N}

\item{colformat}{String (usually indicating format) to add to column (treatment/total) names.}

\item{notrthead}{If no Treatment or subgroup column exists, label to give to "CVALUE" column}
}
\value{
a dataframe with/without selected columns (trt/subgrp) appended with (N = ) values,
or modified column header
}
\description{
Attach bign value to table columns, modify column header labels
}
\examples{
adsl_entry <- mentry(
  datain = adsl,
  trtvar = "TRT01A",
  trtsort = "TRT01AN",
  subgrpvar = "SEX"
)

msumstat(
  adsl_entry,
  dptvar = "AGE",
  statvar = "mean",
  dptvarn = 2
)$tsum |>
  display_bign_head(adsl_entry)
}
