% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edish_plot.R
\name{edish_plot}
\alias{edish_plot}
\title{eDISH Plot}
\usage{
edish_plot(
  datain,
  axis_opts = plot_axis_opts(xlinearopts = list(breaks = c(0.1, 1, 2, 10), limits =
    c(0.1, 10), labels = c("0.1", "1", "2x ULN", "10")), ylinearopts = list(breaks =
    c(0.1, 1, 3, 10), limits = c(0.1, 10), labels = c("0.1", "1", "3x ULN", "10"))),
  xrefline = c("2", "gray30", "dashed"),
  yrefline = c("3", "gray30", "dashed"),
  quad_labels = c("Potential Hy's Law Cases", "Temple's Corollary",
    "Gilberts Syndrome or Cholestasis", "Normal"),
  series_opts = series_opts,
  plot_title = "",
  griddisplay = "N",
  legend_opts = list(label = "Treatment", pos = "bottom", dir = "horizontal"),
  interactive = "N"
)
}
\arguments{
\item{datain}{Input dataset for the plot retreived from \code{process_edish_data()}}

\item{axis_opts}{A \code{list} of axis specific options retrieved from \code{plot_axis_opts()}.}

\item{xrefline}{\code{Xaxis} reference line format.}

\item{yrefline}{\code{Yaxis} reference line format.}

\item{quad_labels}{Labels for each quadrant in the plot, as a vector or a single tilde-separated
string.}

\item{series_opts}{Series Variable styling options, a \code{list} containing
\code{shape}, \code{color} and \code{size}.}

\item{plot_title}{Text to use as plot title, if required}

\item{griddisplay}{Display Grid \code{(Y/N)}.}

\item{legend_opts}{Legend styling option, a \code{list} containing \code{label}, \code{pos}(position) and
\code{dir} (direction).}

\item{interactive}{Interactive plot (\code{'Y'/'N'}).}
}
\value{
A \code{ggplot} or \code{plotly} object
}
\description{
eDISH Plot
}
\details{
\itemize{
\item \code{axis_opts} - The breaks and limits should not be \code{NULL}/empty.
\item \code{quad_labels} - The label for the each quadrant should be populated in the following
order: "upper right~lower right~upper left~lower left" or as a vector of length 4
}
}
\examples{
data("adsl")
data("adlb")

merged_data <- adsl_merge(
  adsl = adsl,
  dataset_add = adlb
) |>
  mentry(
    subset = "SAFFL == 'Y'",
    trtvar = "TRT01A",
    trtsort = "TRT01AN"
  )

pt_data <- process_edish_data(
  datain = merged_data,
  xvar = "both",
  alt_paramcd = "ALT",
  ast_paramcd = "AST",
  bili_paramcd = "BILI"
)

series_opts <- plot_aes_opts(pt_data,
  series_color = NA,
  series_shape = "circlefilled~trianglefilled",
  series_size = c(1.5, 1.5)
)
edish_plot(
  datain = pt_data,
  axis_opts = plot_axis_opts(
    xlinearopts = list(
      breaks = c(0.1, 1, 2, 10),
      limits = c(0.1, 10),
      labels = c("0.1", "1", "2x ULN", "10")
    ),
    ylinearopts = list(
      breaks = c(0.1, 1, 3, 10),
      limits = c(0.1, 10),
      labels = c("0.1", "1", "3x ULN", "10")
    )
  ),
  xrefline = c("2", "gray30", "dashed"),
  yrefline = c("3", "gray30", "dashed"),
  quad_labels =
    "Potential Hy's Law Cases~Temple's Corollary~Gilberts Syndrome or Cholestasis~Normal",
  legend_opts = list(
    label = "Treatment",
    pos = "bottom", dir = "horizontal"
  ),
  series_opts = series_opts,
  plot_title = NULL,
  interactive = "N"
)

}
