% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_display.R
\name{empty_tbl}
\alias{empty_tbl}
\title{Return table if output is empty}
\usage{
empty_tbl(text = "No participant meets the reporting criteria")
}
\arguments{
\item{text}{Text to display under table creation}
}
\value{
flextable output
}
\description{
Return table if output is empty
}
\examples{
empty_tbl()
}
