% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_analysis.R
\name{event_analysis_plot}
\alias{event_analysis_plot}
\title{Event graphs for FDA Medical queries}
\usage{
event_analysis_plot(
  datain,
  fig.align = "h",
  disp.proportion = "3~7",
  ref_line = NA_integer_,
  x_tickangle = 15,
  pt_color = "royalblue3",
  interactive = "N"
)
}
\arguments{
\item{datain}{\code{list} obtained from \code{process_event_analysis}}

\item{fig.align}{Alignment of plots, \code{"h"} for horizontal or \code{"v"} for vertical.}

\item{disp.proportion}{Display proportion of plots horizontally or vertically, should sum up
to \code{10}.}

\item{ref_line}{y axis position of the Horizontal reference line.}

\item{x_tickangle}{angle of \code{x} axis \code{ticks}, default is \code{15}}

\item{pt_color}{Color of bars for Preferred Terms.}

\item{interactive}{Display interactive graph \code{"Y"/"N"}}
}
\value{
\code{ggplot} or \code{plotly} object.
}
\description{
Event graphs for FDA Medical queries
}
\examples{

data(adae)
data(FMQ_Consolidated_List)

## process `ADAE` with `ae_pre_processor()`
prep_ae <- adae |>
  ae_pre_processor(
    ae_filter = "ANY",
    obs_residual = 0,
    fmq_data = FMQ_Consolidated_List
  )

## prepare data for plot
prep_event_analysis <- prep_ae[["data"]] |>
  process_event_analysis(
    a_subset = glue::glue("AOCCPFL == 'Y' & {prep_ae$a_subset}"),
    trtvar = "TRTA",
    trtsort = "TRTAN",
    trttotalyn = "Y",
    summary_by = "Events",
    hterm = "FMQ_NAM",
    ht_val = "ABDOMINAL PAIN",
    ht_scope = "Narrow",
    lterm = "AEDECOD",
    lt_val = "ABDOMINAL DISCOMFORT",
    lt_scope = "Narrow"
  )

## static plot
event_analysis_plot(
  datain = prep_event_analysis,
  disp.proportion = "4~6",
  ref_line = 1
)

## interactive plot
event_analysis_plot(
  datain = prep_event_analysis,
  ref_line = 1,
  interactive = "Y"
)

}
