% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot.R
\name{forest_display}
\alias{forest_display}
\title{Display combined Forest Plot}
\usage{
forest_display(
  plot_list,
  rel_widths = c(0.25, 0.38, 0.27, 0.1),
  interactive = "N",
  plot_height = NULL,
  xpos = "top",
  legend_opts = list(pos = "bottom", dir = "horizontal")
)
}
\arguments{
\item{plot_list}{Named list of ggplot objects to be combined. Must contain the names
"splot" (scatter plot), "fplot" (risk/forest plot) and optionally may include
"ptable" (p-values/other statistic) and/or "termtable" (terms/categories)}

\item{rel_widths}{Relative widths of each subplot - applied to both static and interactive plot}

\item{interactive}{Return interactive subplot object? Alternatively, grid plot returned.
Values: "Y"/"N"}

\item{plot_height}{Height of plotly output, if specifically required}

\item{xpos}{Where should X xaxis for \code{splot} and \code{fplot} be displayed in interactive plot?
Values: "top"/"bottom". Value for static output is decided prior to passing in this function.}

\item{legend_opts}{Legend styling option, a \code{list} containing \code{pos}(position) and
\code{dir} (direction).}
}
\value{
plot_grid object or plotly forest plot object
}
\description{
Display combined Forest Plot
}
\examples{
data(ae_risk)
splot <- forest_plot_scatter(
  datain = ae_risk,
  xvar = "PCT",
  yvar = "DPTVAL",
  series_var = "TRTVAR",
  series_opts = list(
    color = g_seriescol(ae_risk, c("black", "forestgreen"), "TRTVAR"),
    shape = g_seriessym(ae_risk, NA, "TRTVAR"),
    size = rep(1, 2)
  ),
  hovervar = "HOVER_PCT",
  xaxis_pos = "top",
  legend_opts = list(pos = "bottom", dir = "horizontal"),
  axis_opts = list(xsize = 8, xtsize = 6, xaxis_label = "Percentage")
)
fplot <- forest_plot_base(
  ae_risk,
  xvar = "RISK",
  yvar = "DPTVAL",
  xminvar = "RISKCIL",
  xmaxvar = "RISKCIU",
  hovervar = "HOVER_RISK",
  series_var = "TRTPAIR",
  xrefline = 1,
  axis_opts = plot_axis_opts(
    xaxis_label = "Risk Ratio",
    xopts = list(labelsize = 8)
  )
)
forest_display(list(splot = splot, fplot = fplot),
  rel_widths = c(0.6, 0.4)
)
forest_display(list(splot = splot, fplot = fplot),
  rel_widths = c(0.6, 0.4),
  interactive = "Y"
)

}
