% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot.R
\name{forest_plot_base}
\alias{forest_plot_base}
\title{Forest Plot - base with errorbars}
\usage{
forest_plot_base(
  datain,
  xvar = "RISK",
  yvar = "DPTVAL",
  xminvar = "RISKCIL",
  xmaxvar = "RISKCIU",
  hovervar = "HOVER_RISK",
  series_var = "TRTPAIR",
  series_color = c("black", "black"),
  xrefline = 1,
  xaxis_pos = "top",
  legend_opts = list(pos = "bottom", dir = "horizontal"),
  hline_y = "Y",
  axis_opts
)
}
\arguments{
\item{datain}{Input dataset, pre-processed.}

\item{xvar}{X-axis variable (continuous)}

\item{yvar}{Y-axis variable (discrete)}

\item{xminvar}{Min. x value for errorbar (usually lower limit of Confidence Interval)}

\item{xmaxvar}{Max. x value for errorbar (usually upper limit of Confidence Interval)}

\item{hovervar}{Variable containing tooltip info (for interactive plot)}

\item{series_var}{Series Variable (To stratify by).}

\item{series_color}{Vector of colors for series variable}

\item{xrefline}{X-intercept value for vertical reference line}

\item{xaxis_pos}{Position of X axis and title. Values: "top"/"bottom"}

\item{legend_opts}{Legend styling option, a \code{list} containing \code{pos}(position) and
\code{dir} (direction).}

\item{hline_y}{Horizontal line marking each Y variable to be plotted or not? "Y"/"N"}

\item{axis_opts}{A \code{list} of axis specific options, usually retrieved from \code{plot_axis_opts()}.
Following elements - xaxis_label, xaxis_scale, Xlims, Xbrks, xsize, xtsize are utilised and
required from list.}
}
\value{
a ggplot of statistic with interval bars.
}
\description{
Forest Plot - base with errorbars
}
\examples{
data("adae")
ae_pre_process <- ae_pre_processor(
  datain = adae,
  obs_residual = 0
)

ae_entry <- mentry(
  datain = ae_pre_process$data,
  byvar = "AEBODSYS",
  trtvar = "TRTA",
  trtsort = "TRTAN",
  pop_fil = "SAFFL"
)

ae_risk <- risk_stat(
  datain = ae_entry,
  a_subset = ae_pre_process$a_subset,
  summary_by = "Patients",
  eventvar = "AEDECOD",
  ctrlgrp = "Placebo",
  trtgrp = "Xanomeline High Dose",
  statistics = "Risk Ratio",
  alpha = 0.05,
  cutoff_where = "PCT > 2",
  hoveryn = "Y"
) |>
  dplyr::mutate(key = dplyr::row_number())
forest_plot_base(
  ae_risk,
  xvar = "RISK",
  yvar = "DPTVAL",
  xminvar = "RISKCIL",
  xmaxvar = "RISKCIU",
  hovervar = "HOVER_TEXT",
  series_var = "TRTPAIR",
  xrefline = 1,
  axis_opts = plot_axis_opts(
    xaxis_label = "Risk Ratio",
    xopts = list(labelsize = 8)
  )
)
}
