% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot.R
\name{forest_plot_scatter}
\alias{forest_plot_scatter}
\title{Scatter plot to be included within forest plot}
\usage{
forest_plot_scatter(
  datain,
  xvar = "PCT",
  yvar = "DPTVAL",
  series_var = "TRTVAR",
  series_labelvar = series_var,
  series_opts,
  hovervar = "HOVER_PCT",
  xaxis_pos = "top",
  legend_opts = list(pos = "bottom", dir = "horizontal"),
  hline_y = "Y",
  axis_opts = list(xsize = 8, xtsize = 6, xaxis_label = "Percentage")
)
}
\arguments{
\item{datain}{Input dataset, pre-processed.}

\item{xvar}{X-axis variable (continuous)}

\item{yvar}{Y-axis variable (discrete)}

\item{series_var}{Series Variable (To stratify by).}

\item{series_labelvar}{Series (Treatment) Variable labels for legend, if
different from \code{series_var}. If this is needed, then \code{series_labelvar}
should be a factor variable with levels corresponding to \code{series_var}, also a factor.}

\item{series_opts}{Series Variable styling options, a \code{list} containing
\code{shape}, \code{color} and \code{size}.}

\item{hovervar}{Variable containing tooltip info (for interactive plot)}

\item{xaxis_pos}{Position of X axis and title. Values: "top"/"bottom"}

\item{legend_opts}{Legend styling option, a \code{list} containing \code{pos}(position) and
\code{dir} (direction).}

\item{hline_y}{Horizontal line marking each Y variable to be plotted or not? "Y"/"N"}

\item{axis_opts}{A \code{list} of axis specific options.
Only following elements - xaxis_label, xsize, xtsize are utilised and
required from list. eg. list(xaxis_label = "Percentage", xsize = 8, xtsize = 6)}
}
\value{
a ggplot scatterplot, fit for combining with forest plot.
}
\description{
Scatter plot to be included within forest plot
}
\examples{
data("adae")
ae_pre_process <- ae_pre_processor(
  datain = adae,
  obs_residual = 0
)

ae_entry <- mentry(
  datain = ae_pre_process$data,
  byvar = "AEBODSYS",
  trtvar = "TRTA",
  trtsort = "TRTAN",
  pop_fil = "SAFFL"
)

ae_risk <- risk_stat(
  datain = ae_entry,
  a_subset = ae_pre_process$a_subset,
  summary_by = "Patients",
  eventvar = "AEDECOD",
  ctrlgrp = "Placebo",
  trtgrp = "Xanomeline High Dose",
  statistics = "Risk Ratio",
  alpha = 0.05,
  cutoff_where = "PCT > 2",
  hoveryn = "Y"
) |>
dplyr::mutate(key = dplyr::row_number())
forest_plot_scatter(
  datain = ae_risk,
  xvar = "PCT",
  yvar = "DPTVAL",
  series_var = "TRTVAR",
  series_opts = list(
    color = g_seriescol(ae_risk, c("black", "forestgreen"), "TRTVAR"),
    shape = g_seriessym(ae_risk, NA, "TRTVAR"),
    size = rep(1, 2)
  ),
  hovervar = "HOVER_TEXT",
  xaxis_pos = "top"
)

}
