% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{g_seriescol}
\alias{g_seriescol}
\title{Set Colors to Series Variable to plot}
\usage{
g_seriescol(gdata, series_color = NA, SERIESVAR = "TRTVAR")
}
\arguments{
\item{gdata}{Data being used to create graph}

\item{series_color}{Tilde-separated colors corresponding to values in SERIESVAR}

\item{SERIESVAR}{Series Variable to assign colors to. Examples: "TRTVAR","ETHNIC"
Preferably a variable of type factor with distinct levels}
}
\value{
named list of colors per level of grouping variable
}
\description{
Assigns selected colors to levels of a series variable to be used for plotting and legend.
}
\examples{
data("adsl")
g_seriescol(adsl, "red~cyan~forestgreen~black", "TRT01A")
}
