% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{g_seriessym}
\alias{g_seriessym}
\title{Set Shapes to Series Variable to plot}
\usage{
g_seriessym(gdata, series_shape = NA, SERIESVAR = "TRTVAR")
}
\arguments{
\item{gdata}{Data being used to create graph}

\item{series_shape}{Tilde-separated shapes corresponding to values in SERIESVAR}

\item{SERIESVAR}{Series Variable to assign colors to. Examples: "TRTVAR","ETHNIC"
Preferably a variable of type factor with distinct levels}
}
\value{
named list of symbols per level of grouping variable
}
\description{
Assigns selected point shapes to levels of a series variable to be used for plotting and legend.
}
\examples{
data("adsl")
g_seriessym(adsl, "triangle~cross~circle~square", "TRT01A")
}
