% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_plot.R
\name{km_plot}
\alias{km_plot}
\title{Kaplan Meir Plot}
\usage{
km_plot(
  datain,
  disp_conf.int = "N",
  disp_risk.table = "Y",
  risktab_stats = "n.risk",
  risktab_height = NULL,
  disp_pair.stat = "Y",
  trt_colors,
  axis_opts = plot_axis_opts(xlinearopts = list(breaks = 5), ylinearopts = list(breaks =
    0.1), xaxis_label = "Progression-Free Survival Time (Months)", yaxis_label =
    "Probability of Progression Free Survival"),
  legend_opts = list(pos = "bottom", dir = "vertical"),
  time_unit = "Months"
)
}
\arguments{
\item{datain}{Pre-processed input data from \code{surv_pre_processor()}.}

\item{disp_conf.int}{Display confidence interval. Must be one of \code{"Y"} or \code{"N"}.
Default is \code{"N"}.}

\item{disp_risk.table}{Display risk table. Must be one of \code{"Y"} or \code{"N"}. Default is \code{"Y"}.}

\item{risktab_stats}{String of statistics to show in the risk table. Must be one or
more of \code{"n.risk"}, \code{"cum.event"}, \code{"cum.censor"}, \code{"n.event"} and \code{"n.censor"}. Default is
\code{"n.risk"}. See what each option means here \link[ggsurvfit]{add_risktable}).}

\item{risktab_height}{A numeric value between \code{0} and \code{1} to indicate the proportion of final
plot and risk table. Default is \code{NULL}.}

\item{disp_pair.stat}{Display paired prop-hazard statistics.  Must be one of \verb{"Y", "N"}.
Default is \code{"Y"}.}

\item{trt_colors}{String of \code{hex} colors. Must be of the same length of unique \code{TRTVAR}.}

\item{axis_opts}{A \code{list} of axis specific options retrieved from \code{plot_axis_opts()}}

\item{legend_opts}{A \code{list} of legend specific options. Default is
\code{list(pos = "bottom", dir = "vertical")}.}

\item{time_unit}{Unit of \code{timevar} in \code{datain}. Default is \code{"Months"}.}
}
\value{
Kaplan-Meir Plot
}
\description{
Kaplan Meir Plot
}
\examples{
data("survival")

km_df <- survival[["adsl"]] |>
  surv_pre_processor(
    dataset_analysis = survival[["adtte"]],
    analysis_subset = "PARAMCD == 'PFS_P'"
  )

km_df |>
  km_plot(
    trt_colors = "#F8766D~#00BA38~#619CFF"
  )

## with confidence interval and multiple `risk table` statistics
km_df |>
  km_plot(
    disp_conf.int = "Y",
    risktab_stats = "n.risk~n.censor",
    risktab_height = 0.25,
    trt_colors = "#F8766D~#00BA38~#619CFF",
    axis_opts = plot_axis_opts(
      xlinearopts = list(breaks = 3),
      ylinearopts = list(breaks = 0.1),
      xaxis_label = "Progression-Free Survival Time (Months)",
      yaxis_label = "Probability of Progression Free Survival"
    )
  )

## with `{pharmaverseadam}` test data sets
## run `install.packages("pharmaverseadam")` prior running this example
\dontrun{
km_df <- pharmaverseadam::adsl |>
  surv_pre_processor(
    adsl_subset = "SAFFL == 'Y'",
    dataset_analysis = pharmaverseadam::adtte_onco,
    analysis_subset = "PARAMCD == 'OS'",
    trtvar = "TRT01P",
    trtsort = NA_character_,
    censor_val = 0
  )

km_df |>
  km_plot(
    trt_colors = "#F8766D~#00BA38~#619CFF",
    disp_conf.int = "Y",
    time_unit = "Days",
    axis_opts = plot_axis_opts(
      xlinearopts = list(breaks = 25),
      ylinearopts = list(breaks = 0.1),
      xaxis_label = "Overall Survival Time (Days)",
      yaxis_label = "Probability of Overall Survival"
    )
  )
}

}
