% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_abnormality.R
\name{lab_abnormality_summary}
\alias{lab_abnormality_summary}
\title{Incidence of Laboratory Test Abnormalities (Without Regard to Baseline Abnormality)}
\usage{
lab_abnormality_summary(
  datain,
  crit_vars = "CRIT1~CRIT2",
  pctdisp = "SUBGRP",
  a_subset = NA_character_,
  denom_subset = NA_character_,
  sigdec = 2,
  sparseyn = "Y",
  pctsyn = "N",
  stathead = "n (\%)"
)
}
\arguments{
\item{datain}{Input dataset (\code{adlb}).}

\item{crit_vars}{Criteria variables}

\item{pctdisp}{Method to calculate denominator (for \%) by.
Possible values: \code{"TRT"}, \code{"VAR"}, \code{"COL"}, \code{"SUBGRP"}, \code{"CAT"}, \code{"NONE"}, \code{"NO"}, \code{"DPTVAR"},
\code{"BYVARxyN"}}

\item{a_subset}{Analysis Subset condition specific to categorical analysis.}

\item{denom_subset}{Subset condition to be applied to data set for calculating denominator.}

\item{sigdec}{Number of decimal places for \% displayed in output}

\item{sparseyn}{To sparse missing categories/treatments or not? \code{"Y"/"N"}}

\item{pctsyn}{Display Percentage Sign in table or not. Values: \code{"Y"/"N"}}

\item{stathead}{Column label to display \code{n} in the output. Default is \verb{n (\%)}
Values: \verb{"TRT", "VAR","COL", "SUBGRP", "SGRPN", "CAT", "NONE", "NO", "DPTVAR"}}
}
\value{
\code{data.frame} with summary of laboratory abnormality incidence counts
}
\description{
Incidence of Laboratory Test Abnormalities (Without Regard to Baseline Abnormality)
}
\examples{
data("adlb")

lb_entry <- adlb |>
  mentry(
    subset = NA_character_,
    byvar = "PARCAT1~PARAM",
    subgrpvar = NA_character_,
    trtvar = "TRTA",
    trtsort = "TRTAN",
    trttotalyn = "N",
    sgtotalyn = "N",
    add_grpmiss = "N",
    pop_fil = "SAFFL"
  )

out <-
  lb_entry |>
  lab_abnormality_summary(
    crit_vars = "CRIT1~CRIT2",
    pctdisp = "SUBGRP",
    a_subset = NA_character_,
    denom_subset = NA_character_,
    sigdec = 1
  ) |>
  display_bign_head(mentry_data = lb_entry) |>
  tbl_processor()

out

# `flextable` output
out |>
  tbl_display(
    bylabel = "Parameter Category~Parameter",
    dpthead = "Primary Criteria"
  )

}
