% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_plot.R
\name{line_plot}
\alias{line_plot}
\title{Utility for Line Plot}
\usage{
line_plot(
  datain,
  series_var = "TRTVAR",
  series_labelvar = series_var,
  series_opts = plot_aes_opts(datain, "TRTVAR"),
  axis_opts = plot_axis_opts(),
  legend_opts = list(label = "", pos = "bottom", dir = "horizontal"),
  griddisplay = "N",
  plot_title = NULL,
  dodge_width = NULL
)
}
\arguments{
\item{datain}{Input dataset from process_line_plot_data() output.}

\item{series_var}{Series (Treatment) Variable.}

\item{series_labelvar}{Series (Treatment) Variable labels for legend, if
different from \code{series_var}. If this is needed, then \code{series_labelvar}
should be a factor variable with levels corresponding to \code{series_var}, also a factor.}

\item{series_opts}{Series Variable styling options, a \code{list} containing
\code{shape}, \code{color} and \code{size}.}

\item{axis_opts}{A \code{list} of axis specific options retrieved from \code{plot_axis_opts()}.}

\item{legend_opts}{Legend styling option, a \code{list} containing \code{label}, \code{pos}(position) and
\code{dir} (direction).}

\item{griddisplay}{Display Grid \code{(Y/N)}.}

\item{plot_title}{Text to use as plot title, if required}

\item{dodge_width}{Width to dodge points/lines by, IF required.}
}
\value{
plot - Line plot.
}
\description{
Utility for Line Plot
}
\examples{
data(adsl)

mentry_df <- adsl |>
  mentry(
    subset = NA_character_,
    byvar = NA_character_,
    trtvar = "TRT01A",
    trtsort = "TRT01AN",
    subgrpvar = NA_character_,
    trttotalyn = "N",
    add_grpmiss = "N",
    pop_fil = "SAFFL"
  )
plot_df <- mcatstat(
  mentry_df,
  dptvar = "RACE"
) |>
  dplyr::mutate(YVAR = PCT)
line_plot(
  datain = plot_df,
  series_var = "TRTVAR",
  series_labelvar = "TRTVAR",
  series_opts = plot_aes_opts(
  plot_df, "TRTVAR", series_color = "firebrick~forestgreen~dodgerblue",
  series_shape = "triangle~square~circle"),
  axis_opts = plot_axis_opts(
    ylinearopts = list(
      limits = c(0, 100),
      breaks = seq(0, 100, by = 10)
    ),
    xaxis_label = "Race",
    yaxis_label = "Percent of Subjects"
  ),
  legend_opts = list(
    label = "",
    pos = "bottom",
    dir = "horizontal"
  ),
  griddisplay = "N",
  plot_title = "Line Plot of Subjects by Race"
)

}
