% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mentry.R
\name{mentry}
\alias{mentry}
\title{Read and process data with subsets and variables}
\usage{
mentry(
  datain,
  subset = NA,
  byvar = NA,
  subgrpvar = NA,
  trtvar = NA,
  trtsort = NA,
  trttotalyn = "N",
  trttotlabel = "Total",
  trtmissyn = "N",
  sgtotalyn = "N",
  add_grpmiss = "N",
  pop_fil = "Overall Population"
)
}
\arguments{
\item{datain}{Input dataset can be any \code{ADaM} data.}

\item{subset}{Subset conditions for analysis of dependent variable.}

\item{byvar}{Tilde(\code{~}) separated strings for grouping variable names. Assigns to
\verb{BYVAR1, BYVAR2}, etc. Use \code{'/'} to specify ordering variables. eg: \code{SEX/SEXN~COUNTRY}.`}

\item{subgrpvar}{Tilde(\code{~}) separated strings for grouping variable names. Assigns to
\code{SUBGRPVAR1}, \code{SUBGRPVAR2}, etc. Use \code{'/'} to specify ordering variables.
eg: \code{RACE/RACEN~ETHNIC}.}

\item{trtvar}{Treatment variable name string assigned as \code{TRTVAR}.}

\item{trtsort}{Treatment sorting variable name string - numeric or
character variable assigned as \code{TRTSORT}.}

\item{trttotalyn}{Add total treatment values to be displayed as column in
table or category in plot (\code{"Y"/"N"}).}

\item{trttotlabel}{Label for total Treatment column/group}

\item{trtmissyn}{Retain Missing treatment counts in Total (if \code{trttotalyn} = Y). Missing
treatment will not be considered as a column in analysis in any case.}

\item{sgtotalyn}{Add total subgroup values to be displayed as column in
table or category in plot (\code{"Y"/"N"}).}

\item{add_grpmiss}{Add row or column for missing category in grouping
variables. (\code{"Y"/"N"})}

\item{pop_fil}{Population Filter for data set: Name of flag variable.
eg: \code{"SAFFL"}, \code{"EFFFL"} or \code{NA} for Overall Population.}
}
\value{
a \code{data.frame} output to be used in further utilities with applied \code{subset}
}
\description{
The input data gets processed based on \code{TRT}, \code{BY} and/or \code{SUBGROUP} variables and filtered by
subset conditions. This processed data flows into categorical and numerical utilities such as
\code{mcatstat()} and \code{msumstat()} for further analysis.
}
\details{
\enumerate{
\item \strong{By Variable Processing}
Identify \code{byvars} if byvar is given, create \code{BYVAR1}, \code{BYVAR2}, etc.
Check for numeric code list variable for the grouping var, if not
create numeric variable for sorting as \code{BYVAR1N}, \code{BYVAR2N} etc.
Convert to \code{character} if \code{numeric}.
Re-code blanks from grouping variables as \code{"Missing"}
Remove missing category value from grouping variable per user choice.
\item \strong{Treatment Variable Processing}
Identify Treatment Variable, treatment sorting variable.
Remove non treatment values.
If treatment sort variable doesn't exist, make it same as \code{TRTVAR}
If treatment sort is not numeric, convert to numeric
Add treatment total column based on \code{trttotalyn}
Convert \code{TRTVAR} to a factor in same order as \code{TRTSORT}
\item \strong{Subgroup Variable processing}
Identify Subgroup variable: create \code{SUBGRPVAR1N}, \code{SUBGRPVAR2N} etc if \code{subgrpvar} is
specified.
Check presence of numeric code list variables for the subgroup variables,
if not create numeric variable for sorting purpose as \code{SUBGRPVAR1N} etc
Re-code blanks from By and Subgroup vars as \code{Missing}
Remove missing category value from grouping variable per user choice
Add Subgroup total column for the last subgroup variable
\item \strong{Subset Processing}
Applying population subset selected
Applying denominator/overall subset condition passed by the user
Applying analysis subset condition passed by the user
\item \strong{Calculate Big N for displaying in final output}
\code{data.frame} with Count \code{N} for each treatment/subgroup if they exist.
If no \verb{Big N} required, then \code{NA} is returned.
}
}
\examples{
data(adsl)

df_mentry <-
  adsl |> mentry(
    subset = "EFFFL=='Y'",
    byvar = "SEX/SEXN~AGEGR1",
    trtvar = "TRT01A",
    trtsort = "TRT01AN",
    subgrpvar = "RACE/RACEN~BMIBLGR1",
    trttotalyn = "N",
    add_grpmiss = "N",
    sgtotalyn = "N",
    pop_fil = "Overall Population"
  )

df_mentry
}
