% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_tier_summary.R
\name{occ_tier_summary}
\alias{occ_tier_summary}
\title{Generic Occurrence Summary Tiered Table}
\usage{
occ_tier_summary(
  datain,
  a_subset = NA_character_,
  summary_by = "Patients",
  hterm = "AEBODSYS",
  lterm = "AEDECOD",
  pctdisp = "TRT",
  cutoff = 2,
  apply_hrow_cutoff = "N",
  sort_opt = "Ascending",
  sort_var = "Count"
)
}
\arguments{
\item{datain}{Input dataset (generally the output from \code{mentry()})}

\item{a_subset}{Analysis Subset condition specific to categorical analysis.}

\item{summary_by}{Measure to construct the summary by. Values: \verb{'Patients' or 'Events'}.}

\item{hterm}{High Level Event term variable, used for analysis}

\item{lterm}{Low Level Event term variable, used for analysis}

\item{pctdisp}{Method to calculate denominator (for \%) by.
Possible values: \code{"TRT"}, \code{"VAR"}, \code{"COL"}, \code{"SUBGRP"}, \code{"CAT"}, \code{"NONE"}, \code{"NO"}, \code{"DPTVAR"},
\code{"BYVARxyN"}}

\item{cutoff}{Incidence Cutoff Value; consider only terms with \verb{incidence percentage > cutoff}.}

\item{apply_hrow_cutoff}{To apply cutoff value to high terms in addition to low term.
If set to "Y" same cutoff is applied to remove both  high and low level terms that don't meet
the criteria.
If set to "N" (default), cutoff is applied only to Lower Level term. The terms that do not fit
the criteria are then excluded from the counts for High Level term. This does not happen in case
of "N" - all counts are included in high term which is displayed as long as it meets the criteria
as well.}

\item{sort_opt}{How to sort terms, only for table/forest plot.
Values: \verb{'Ascending','Descending','Alphabetical'}.}

\item{sort_var}{Metric to sort by. Values: \verb{'Count','Percent','RiskValue'}.}
}
\value{
Summarized data frame for Adverse Events based on high and lower terms.
}
\description{
Generic Occurrence Summary Tiered Table
}
\examples{
ae_entry <- ae_pre_process[["data"]] |>
  mentry(
    subset = NA,
    byvar = "AEBODSYS",
    trtvar = "TRTA",
    trtsort = "TRTAN",
    trttotalyn = "N",
    add_grpmiss = "N",
    sgtotalyn = "N",
    pop_fil = "Overall Population"
  )
output <- occ_tier_summary(
  ae_entry,
  a_subset = ae_pre_process[["a_subset"]],
  summary_by = "Patients",
  hterm = "AEBODSYS",
  lterm = "AEDECOD",
  pctdisp = "TRT",
  cutoff = 2,
  apply_hrow_cutoff = "N",
  sort_opt = "Ascending",
  sort_var = "Count"
)
output |>
  tbl_processor() |>
  tbl_display()

}
