% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph_plot.R
\name{ph_plot}
\alias{ph_plot}
\title{Function to create Proportional Hazard Plot}
\usage{
ph_plot(
  datain,
  series_opts,
  pair_id = NA,
  axis_opts = plot_axis_opts(),
  legend_opts = list(pos = "bottom", label = "Treatment Group:"),
  griddisplay = "N"
)
}
\arguments{
\item{datain}{Input \code{data.frame}.}

\item{series_opts}{Series Variable styling options, a \code{list} containing
\code{shape}, \code{color} and \code{size}.}

\item{pair_id}{Pairs for which pairwise statistics are generated in the form
1-2
Default: NA - which displays all the combination.
Valid Values: specify the pairs between '-' and each pair should be delimited
by \code{~}.
Example: "1-2~2-3"}

\item{axis_opts}{A \code{list} of axis specific options retrieved from \code{plot_axis_opts()}.}

\item{legend_opts}{Legend styling option, a \code{list} containing \code{label}, \code{pos}(position) and
\code{dir} (direction).}

\item{griddisplay}{Display Grid \code{(Y/N)}.}
}
\value{
proportional hazard plot for the each pair of treatments.
}
\description{
Function to create Proportional Hazard Plot
}
\examples{
data("survival")

survival$adsl <- survival$adsl |>
  dplyr::mutate(
    TRT01PN = dplyr::case_when(
      TRT01P == "Xanomeline Low Dose" ~ 1,
      TRT01P == "Placebo" ~ 2,
      TRT01P == "Screen Failure" ~ 3,
      TRUE ~ NA_real_
    )
  )

ph_pre <- surv_pre_processor(
  dataset_adsl = survival$adsl,
  adsl_subset = "SAFFL=='Y'",
  dataset_analysis = survival$adtte,
  split_by = NA_character_,
  analysis_subset = "PARAMCD=='PFS'",
  trtsort = "TRT01PN",
  censor_var = "CNSR",
  censor_val = 1,
  trtvar = "TRT01P",
  time_var = "AVAL"
)

series_opts <- plot_aes_opts(ph_pre,
  series_color = "red~blue~green",
  series_shape = c(1, 1, 1),
  series_size = c(1.5, 1.5, 1.5)
)

ph_plot(
  datain = ph_pre,
  axis_opts = plot_axis_opts(),
  series_opts = series_opts,
  pair_id = "1-2"
)

}
