% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{plot_aes_opts}
\alias{plot_aes_opts}
\title{Set Series Variable Options for color, shape and size}
\usage{
plot_aes_opts(
  datain,
  series_var = "TRTVAR",
  series_color = NA,
  series_shape = NA,
  series_size = rep(1, 10),
  series_contrast = ""
)
}
\arguments{
\item{datain}{Input dataset to be plotted}

\item{series_var}{Series Variable}

\item{series_color}{Tilde-separated string or vector of colors to be set to
levels of \code{series_var}. Pass NA to get package defaults}

\item{series_shape}{Tilde-separated string or vector of symbols to be set to
levels of \code{series_var}. Pass NA to get package defaults}

\item{series_size}{Tilde-separated string or vector of sizes to be set to
levels of \code{series_var}. Pass NA to get package defaults}

\item{series_contrast}{Tilde-separated string or vector of colors to be set to
levels of \code{series_var}. (Optional)}
}
\value{
a list containing vectors for colors, shapes and sizes matched to levels in
Series Variable
}
\description{
Set Series Variable Options for color, shape and size
}
\details{
The \code{series_shape} argument can take numeric values such as
c(16, 25, 22), "16,25,22" or characters -
c("triangle", "square")/"triangle,square"
}
\examples{
plot_aes_opts(iris,
  series_var = "Species",
  series_color = NA,
  series_shape = "trianglefilled~circlefilled~squarefilled~asterisk",
  series_size = c(2, 2, 2, 2),
  series_contrast = "black~black~black"
)
}
