% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{plot_axis_opts}
\alias{plot_axis_opts}
\title{Plot Axes Options}
\usage{
plot_axis_opts(
  xlinearopts = list(breaks = waiver(), limits = NULL, labels = waiver()),
  ylinearopts = list(breaks = waiver(), limits = NULL, labels = waiver()),
  xaxis_scale = "identity",
  yaxis_scale = "identity",
  xaxis_label = "",
  yaxis_label = "",
  xopts = list(labelsize = 12, labelface = "plain", ticksize = 8, tickface = "plain",
    angle = 0),
  yopts = list(labelsize = 12, labelface = "plain", ticksize = 8, tickface = "plain",
    angle = 0)
)
}
\arguments{
\item{xlinearopts}{Linear X axis options for axis breaks and limits.}

\item{ylinearopts}{Linear Y axis options for axis breaks and limits}

\item{xaxis_scale}{Scaletype to transform X axis of plot}

\item{yaxis_scale}{Scaletype to transform Y axis of plot}

\item{xaxis_label}{Label for X axis}

\item{yaxis_label}{Label for Y axis}

\item{xopts}{Options for X axis labels}

\item{yopts}{Options for Y axis labels}
}
\value{
Combined list of X and Y axis options
}
\description{
Plot Axes Options
}
\examples{
plot_axis_opts(
  xlinearopts = list(
    breaks = c(0.001, 0.01, 0.1, 1, 10, 100),
    limits = c(0.001, 100)
  ),
  ylinearopts = list(
    breaks = c(0.001, 0.01, 0.1, 1, 10, 100, 1000),
    limits = c(0.001, 1000)
  )
)

plot_axis_opts(
  xlinearopts = list(
    breaks = c(0.001, 0.01, 0.1, 1, 10, 100),
    limits = c(0.001, 100)
  ),
  ylinearopts = list(
    breaks = c(0.001, 0.01, 0.1, 1, 10, 100, 1000),
    limits = c(0.001, 1000)
  ),
  xopts = list(
    labelsize = 12,
    labelface = "plain",
    ticksize = 8,
    tickface = "plain"
  ),
  yopts = list(
    labelsize = 15,
    labelface = "plain",
    ticksize = 10,
    tickface = "plain"
  )
)

}
