% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{plot_display_bign}
\alias{plot_display_bign}
\title{Return N count in plot legend (treatment)}
\usage{
plot_display_bign(datain, mentry_data, bignyn = "Y")
}
\arguments{
\item{datain}{Input dataframe for display}

\item{mentry_data}{Input data from \code{mentry()}}

\item{bignyn}{Display count "(N = )" along with Treatment in plot legend? Values: "Y"/"N"}
}
\value{
dataframe with additional variable "TRTTXT" containing the label for Treatment
}
\description{
Return N count in plot legend (treatment)
}
\examples{
data(adsl)
adsl_entry <- adsl |>
  mentry(
    trtvar = "TRT01A",
    trtsort = "TRT01AN",
    trttotalyn = "N",
    add_grpmiss = "N",
    pop_fil = "SAFFL"
  )
msumstat(
  adsl_entry,
  dptvar = "AGE",
  statvar = "meansd"
)$gsum |>
  plot_display_bign(adsl_entry)
}
