% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptly_utils.R
\name{plotly_legend}
\alias{plotly_legend}
\title{Legend Positioning and Labeling for an interactive Plot}
\usage{
plotly_legend(fig, lg_pos = "bottom", lg_lab = "", dir = "h")
}
\arguments{
\item{fig}{Plotly or subplot object as input (with ggplot legend position none)}

\item{lg_pos}{Legend Position. Values: "right","left","top","bottom", c(1,2), c(5,7) etc}

\item{lg_lab}{Legend Title Text; if any.}

\item{dir}{Direction of legend. Values: "h", "v", "horizontal", "vertical"}
}
\value{
Plotly object with legend positioned and titled according to input
}
\description{
By default, Plotly only creates legend on the right - use to modify position and label
accordingly
}
\examples{
library(ggplot2)

g <- ggplot(iris, aes(x = Petal.Length, y = Petal.Width, color = Species)) +
  geom_point() +
  theme(legend.position = "none")
plotly_legend(plotly::ggplotly(g), "bottom", "Species:")
}
