% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edish_plot.R
\name{process_edish_data}
\alias{process_edish_data}
\title{Process data for eDISH plot}
\usage{
process_edish_data(
  datain,
  xvar = "both",
  alt_paramcd = "L00030S",
  ast_paramcd = "L00028S",
  bili_paramcd = "L00021S"
)
}
\arguments{
\item{datain}{Input dataset.}

\item{xvar}{To determine how \code{ast} and \code{alt} should be derived and which should be the
\code{x} variable for the plot. Possible values are: \code{"both"}, \code{"ast"}, \code{"alt"}.
If it is "ast" or "alt", the \code{AVAL} for the corresponding "PARAMCD" is used as X variable.
Alternatively, if "both" then the max value between ast and alt for each record is used as X
variable.}

\item{alt_paramcd}{\code{PARAMCD} value for \verb{ALANINE AMINOTRANSFERASE} in \code{datain}.}

\item{ast_paramcd}{\code{PARAMCD} value for \verb{ASPARTATE AMINOTRANSFERASE} in \code{datain}.}

\item{bili_paramcd}{\code{PARAMCD} value for \code{BILIRUBIN} in \code{datain}.}
}
\value{
A \code{data.frame} required for \code{edish_plot}.
}
\description{
Process data for eDISH plot
}
\examples{
data("adlb")
data("adsl")

merged_data <- adsl_merge(
  adsl = adsl,
  dataset_add = adlb
) |>
  mentry(
    subset = "SAFFL == 'Y'",
    trtvar = "TRT01A",
    trtsort = "TRT01AN"
  )

merged_data |>
  process_edish_data(
    xvar = "both",
    alt_paramcd = "ALT",
    ast_paramcd = "AST",
    bili_paramcd = "BILI"
  )

}
