% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_analysis.R
\name{process_event_analysis}
\alias{process_event_analysis}
\title{Process data for Event Analysis}
\usage{
process_event_analysis(
  datain,
  a_subset = NA_character_,
  trtvar,
  trtsort = NA_character_,
  trttotalyn = "N",
  summary_by = "Events",
  hterm,
  ht_val,
  ht_scope,
  lterm,
  lt_val,
  lt_scope
)
}
\arguments{
\item{datain}{Input dataset can be any \code{ADaM} data.}

\item{a_subset}{Analysis subset condition.}

\item{trtvar}{Treatment variable name string assigned as \code{TRTVAR}.}

\item{trtsort}{Treatment sorting variable name string - numeric or
character variable assigned as \code{TRTSORT}.}

\item{trttotalyn}{Add total treatment values to be displayed as column in
table or category in plot (\code{"Y"/"N"}).}

\item{summary_by}{Set to \code{'Events'} or \code{'Patients'}.}

\item{hterm}{\code{MedDRA} queries or high level variables; valid values : \code{"FMQ_NAM"}, \code{"AEBODSYS"}}

\item{ht_val}{Queries/term name captured in the \code{hl_var} variable, e.g.\code{"Erythema"}}

\item{ht_scope}{Scope of the \code{MedDRA} queries; Valid values : \code{"Narrow"} or \code{"Broad"}}

\item{lterm}{\code{MedDRA} variable; Valid value : \code{"AEDECOD"}}

\item{lt_val}{\code{MedDRA} value, e.g., \code{"Erythema"}}

\item{lt_scope}{Scope of the \code{MedDRA} queries; Valid values : \code{"Narrow"} or \code{"Broad"}}
}
\value{
List of data frames summarized by \code{hterm} and \code{lterm}
}
\description{
Process data for Event Analysis
}
\examples{
data(adae)
data(FMQ_Consolidated_List)

## process `ADAE` with `ae_pre_processor()`
prep_ae <- adae |>
  ae_pre_processor(
    ae_filter = "ANY",
    obs_residual = 0,
    fmq_data = FMQ_Consolidated_List
  )

## prepare data for plot
prep_event_analysis <- prep_ae[["data"]] |>
  process_event_analysis(
    a_subset = glue::glue("AOCCPFL == 'Y' & {prep_ae$a_subset}"),
    trtvar = "TRTA",
    trtsort = "TRTAN",
    trttotalyn = "Y",
    summary_by = "Events",
    hterm = "FMQ_NAM",
    ht_val = "ABDOMINAL PAIN",
    ht_scope = "Narrow",
    lterm = "AEDECOD",
    lt_val = "ABDOMINAL DISCOMFORT",
    lt_scope = "Narrow"
  )

prep_event_analysis[["pt_df"]]
prep_event_analysis[["query_df"]]

}
