% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_vx_bar_plot.R
\name{process_vx_bar_plot}
\alias{process_vx_bar_plot}
\title{Pre-Process data for Bar Plot}
\usage{
process_vx_bar_plot(
  dataset_adsl,
  adsl_subset = "SAFFL=='Y'",
  dataset_analysis,
  analysis_subset = NA_character_,
  overall_subset = NA_character_,
  denom_subset = NA_character_,
  split_by = NA_character_,
  trtvar = "TRT01A",
  trtsort = "TRT01AN",
  xvar = "ATPTN",
  yvar = "PCT",
  pctdisp = "DPTVAR",
  legendbign = "Y"
)
}
\arguments{
\item{dataset_adsl}{(\code{data.frame}) ADSL dataset.}

\item{adsl_subset}{(\code{string}) Subset condition to be applied on \code{dataset_adsl}.}

\item{dataset_analysis}{(\code{data.frame}) Analysis Dataset.}

\item{analysis_subset}{Subset conditions for analysis of dependent variable
Applicable only to numerator calculation for \%}

\item{overall_subset}{Subset conditions for overall data.}

\item{denom_subset}{Subset condition to be applied to data set for
calculating denominator.}

\item{split_by}{(\code{string}) By variable for stratification.}

\item{trtvar}{(\code{string}) Treatment Variable to be created for analysis.}

\item{trtsort}{(\code{string}) Variable to sort treatment variable by.}

\item{xvar}{Categorical Analysis variable for X axis}

\item{yvar}{Y axis variable/statistic. Possible Values: "FREQ"/"PCT"}

\item{pctdisp}{Method to calculate denominator (for \%) by
Possible values: "TRT","VAR","COL","SUBGRP","CAT","NONE","NO","DPTVAR"}

\item{legendbign}{(\code{string}) Display count as (N = ..) in Treatment legend? Values: "Y"/"N"}
}
\value{
mcatstat dataset as data frame.
}
\description{
Pre-Process data for Bar Plot
}
\details{
\itemize{
\item Subset Processing
Applying population subset selected
Applying denominator/overall subset condition passed by the user
Applying analysis/numerator subset condition passed by the user.
\item pctdisp has possible values for method to get denominator to calculate
percentage, passed to \code{mcatstat()}. The commonly passed value for vaccine
bar plot is:
DPTVAR: Percentage within each Treatment-By group(s)-Subgroup(s)-dptvar
combination.
}
}
\examples{
data(vx_bar_data)

process_vx_bar_plot(
  dataset_adsl = vx_bar_data$adsl,
  adsl_subset = "SAFFL=='Y'",
  dataset_analysis = vx_bar_data$adfacevd,
  analysis_subset = "ATPTN <= 14 & toupper(FAOBJ) == 'PAIN AT INJECTION SITE' &
 !(AVAL \%in\% c(0, 0.5)) & FATESTCD != 'OCCUR' & !is.na(AVAL)",
  denom_subset = "ATPTN <= 14 & toupper(FAOBJ) == 'PAIN AT INJECTION SITE' &
 !(AVAL \%in\% c(0, 0.5))",
  overall_subset = NA,
  split_by = "SEX",
  trtvar = "TRT01A",
  trtsort = "TRT01AN",
  xvar = "ATPTN",
  yvar = "PCT",
  pctdisp = "DPTVAR"
)

}
