% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_plot.R
\name{process_vx_scatter_data}
\alias{process_vx_scatter_data}
\title{Process data for Vaccines Scatter Plot}
\usage{
process_vx_scatter_data(
  dataset_adsl,
  dataset_analysis,
  adsl_subset,
  analysis_subset = NA_character_,
  split_by = NA_character_,
  trtvar,
  trtsort = NA,
  xvar = "AVISITN == 1",
  yvar = "AVISITN == 2",
  legendbign = "Y"
)
}
\arguments{
\item{dataset_adsl}{(\code{data.frame}) ADSL dataset.}

\item{dataset_analysis}{(\code{data.frame}) Analysis Dataset.}

\item{adsl_subset}{(\code{string}) Subset condition to be applied on \code{dataset_adsl}.}

\item{analysis_subset}{(\code{string}) Subset Condition to be applied on \code{dataset_analysis}.}

\item{split_by}{(\code{string}) By variable for stratification.}

\item{trtvar}{(\code{string}) Treatment Variable to be created for analysis.}

\item{trtsort}{(\code{string}) Variable to sort treatment variable by.}

\item{xvar}{(\code{string}) Values for X axis, determined by filter condition for
analysis visit.}

\item{yvar}{(\code{string}) Values for Y axis, determined by filter condition for
analysis visit.}

\item{legendbign}{(\code{string}) Display count as (N = ..) in Treatment legend? Values: "Y"/"N"}
}
\value{
Grouped Data Frames within a list
}
\description{
Process data for Vaccines Scatter Plot
}
\examples{
data("vx_scatter_data")

process_vx_scatter_data(
  dataset_adsl = vx_scatter_data[["adsl"]],
  dataset_analysis = vx_scatter_data[["adva"]],
  adsl_subset = 'EVALFL=="Y"',
  analysis_subset = 'ANL01FL=="Y" & PARAMN==23 &
((AVISITN ==1 & EVALFL=="Y")|(AVISITN==2 & EVALFL=="Y" & EXCL5FL=="N")) &
 TRTA!="" & DTYPE=="LLOQIMP" & !is.na(AVAL)', # nolint
  split_by = "SEX",
  trtvar = "ACTARM",
  xvar = "AVISITN == 1",
  yvar = "AVISITN == 2"
)

}
