% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_f}
\alias{round_f}
\title{Round and Format values}
\usage{
round_f(x, digits = 2)
}
\arguments{
\item{x}{Numeric value to be rounded}

\item{digits}{Number of decimal places}
}
\value{
Rounded and formatted character value
}
\description{
Rounds value and formats it to retain number of decimals in 2nd argument
}
\examples{
round_f(12.31, 3)
}
