% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{shape_to_sym}
\alias{shape_to_sym}
\title{Recode Shapes to Numbers}
\usage{
shape_to_sym(shape)
}
\arguments{
\item{shape}{Vector of valid shapes}
}
\value{
Vector of numbers
}
\description{
Recode Shapes to Numbers
}
\examples{
shape_to_sym("triangle")
}
