% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_data_by_var}
\alias{split_data_by_var}
\title{Split data by given variables into list}
\usage{
split_data_by_var(datain, split_by = "", split_by_prefix = "")
}
\arguments{
\item{datain}{Input data}

\item{split_by}{Variables to split data by (used if \code{split_by_prefix} is "")}

\item{split_by_prefix}{Prefix to identify splitting variables}
}
\value{
list of dataframe(s)
}
\description{
Split data by given variables into list
}
\examples{
split_data_by_var(iris, "Species")
}
