% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_section_headers}
\alias{split_section_headers}
\title{Return section headers according to splitting variables specified}
\usage{
split_section_headers(
  datain,
  split_by = "",
  split_by_prefix = "",
  split_lab = " ",
  sep = "~"
)
}
\arguments{
\item{datain}{Input data}

\item{split_by}{Variables to split data by (used if \code{split_by_prefix} is "")}

\item{split_by_prefix}{Prefix to identify splitting variables}

\item{split_lab}{Text to prefix the section values with}

\item{sep}{Separating character if multiple split_by are present}
}
\value{
list of dataframe(s)
}
\description{
Return section headers according to splitting variables specified
}
\examples{

library(dplyr)
data(adsl)

adsl |>
  mutate(SUBGRPVAR1 = SEX, SUBGRPVAR2 = AGEGR1) |>
  split_section_headers(
    split_by = "",
    split_by_prefix = "SUBGRPVAR",
    split_lab = "Group:~Age Group",
    sep = " & "
  )

iris |>
  split_section_headers("Species", split_lab = "Name of species:")
}
