% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_to_vec}
\alias{str_to_vec}
\title{Convert string to vector}
\usage{
str_to_vec(var, sep = "~")
}
\arguments{
\item{var}{The string object}

\item{sep}{Separator between strings. Default: "~"}
}
\value{
Character vector
}
\description{
Convert string to vector
}
