% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_utils.R
\name{summary_functions}
\alias{summary_functions}
\title{List of Summary Functions}
\usage{
summary_functions(statvar, statdec)
}
\arguments{
\item{statvar}{Input statistics}

\item{statdec}{Corresponding number of decimal places for each statistic}
}
\value{
A named list containing function definition for all defined summary
statistics - mean, min, max, median, mode iqr, var, sum, sd, q25, q75, p1, p5,  # nolint
p10, p90, p95, p99 (where last digits represent \% of quantile), whiskerlow,
whiskerup, outliers in the Tukey method for box statistics, geometric mean/sd/CI  # nolint
}
\description{
List of Summary Functions
}
\examples{
summary_functions(c("mean", "mode"), c(2, 1))
}
