% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_tier_summary.R
\name{summary_row_cat}
\alias{summary_row_cat}
\title{Insert Overall/Summary Row}
\usage{
summary_row_cat(
  datain,
  sum_row_label = "Any Term",
  byvaryn = "N",
  a_subset = NA,
  pctdisp = "TRT",
  sigdec = 2,
  pctsyn = "Y",
  var = "ANY",
  uniqid = "USUBJID"
)
}
\arguments{
\item{datain}{Input dataset \code{ADAM} or intermediate within summary function}

\item{sum_row_label}{Label for Summary Row to be displayed, if Y.}

\item{byvaryn}{Include by variable or not? For single overally row, "N"}

\item{a_subset}{Analysis Subset condition specific to categorical analysis.}

\item{pctdisp}{Method to calculate denominator (for \%) by.
Possible values: \code{"TRT"}, \code{"VAR"}, \code{"COL"}, \code{"SUBGRP"}, \code{"CAT"}, \code{"NONE"}, \code{"NO"}, \code{"DPTVAR"},
\code{"BYVARxyN"}}

\item{sigdec}{Number of decimal places for \% displayed in output}

\item{pctsyn}{Display Percentage Sign in table or not. Values: \code{"Y"/"N"}}

\item{var}{Flag Variable to identify Any/Summary Rows}

\item{uniqid}{Variable(s) to calculate unique counts of. eg. \code{"USUBJID"}, \code{"SITEID"},
\code{"ALLCT"}}
}
\value{
dataframe with single overall row count
}
\description{
Insert Overall/Summary Row
}
\examples{
data("adae")
summary_row_cat(
  adae,
  a_subset = "TRTEMFL == 'Y'"
)

}
