% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_utils.R
\name{surv_pre_processor}
\alias{surv_pre_processor}
\title{Process data for Survival Analysis}
\usage{
surv_pre_processor(
  dataset_adsl,
  dataset_analysis,
  adsl_subset = "SAFFL=='Y'",
  analysis_subset = NA_character_,
  split_by = NA_character_,
  trtvar = "TRT01P",
  trtsort = "TRT01PN",
  censor_var = "CNSR",
  censor_val = 1,
  time_var = "AVAL"
)
}
\arguments{
\item{dataset_adsl}{(data.frame) ADSL dataset.}

\item{dataset_analysis}{(data.frame) Analysis Dataset.}

\item{adsl_subset}{(string) Subset condition to be applied on dataset_adsl.}

\item{analysis_subset}{(string) Subset Condition to be applied on dataset_analysis.}

\item{split_by}{(string) By variable for stratification.}

\item{trtvar}{(string) Treatment Variable to be created for analysis.}

\item{trtsort}{(string) Variable to sort treatment variable by.}

\item{censor_var}{Censoring Variable in the input dataset to be used in
ph reg model.Default: \code{"CNSR"}}

\item{censor_val}{Value within \code{CNSR} variable to be considered as Censor
value.Default: \code{1}}

\item{time_var}{Duration variable in the input dataset to be used in
proportional hazard regression model. Default: \code{"AVAL"}}
}
\value{
Data frame with added variables for survival analysis
}
\description{
Process data for Survival Analysis
}
\examples{
data("survival")

survival$adsl |>
  surv_pre_processor(
    dataset_analysis = survival$adtte,
    adsl_subset = "SAFFL=='Y'",
    analysis_subset = "PARAMCD=='OS' & FASFL=='Y'",
    split_by = NA_character_,
    trtvar = "TRT01P",
    trtsort = "TRT01PN",
    censor_var = "CNSR",
    censor_val = 1,
    time_var = "AVAL"
  )
}
