% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph_plot.R
\name{survfit_summary}
\alias{survfit_summary}
\title{Customized function to get the summary dataframe from survfit()}
\usage{
survfit_summary(x, data)
}
\arguments{
\item{x}{Input list object created from survfit().}

\item{data}{Pair wise dataframe.}
}
\value{
The summary dataframe with cleaned TRTVAR variable.
}
\description{
Customized function to get the summary dataframe from survfit()
}
\examples{
data("survival")
survival$adsl <- survival$adsl |>
  dplyr::mutate(
    TRT01PN = dplyr::case_when(
      TRT01P == "Xanomeline Low Dose" ~ 1,
      TRT01P == "Placebo" ~ 2,
      TRT01P == "Screen Failure" ~ 3,
      TRUE ~ NA_real_
    )
  )
ph_pre <- surv_pre_processor(
  dataset_adsl = survival$adsl,
  adsl_subset = "SAFFL=='Y'",
  dataset_analysis = survival$adtte,
  split_by = NA_character_,
  analysis_subset = "PARAMCD=='PFS'",
  trtsort = "TRT01PN",
  censor_var = "CNSR",
  censor_val = 1,
  trtvar = "TRT01P",
  time_var = "AVAL"
)
survfit_df <- survival::survfit(survival::Surv(timevar, cnsrvar) ~ TRTVAR,
data = ph_pre)
survfit_summary(
  x = survfit_df,
  data = pair_data
)
}
