% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_display.R
\name{tbl_display}
\alias{tbl_display}
\title{Create flextable output from display templates}
\usage{
tbl_display(
  datain,
  bylabel = NA,
  dpthead = "  ",
  font = "Arial",
  fontsize = 10,
  boldheadyn = "N"
)
}
\arguments{
\item{datain}{Input dataframe}

\item{bylabel}{Change \code{BYVAR} names to label, if any.}

\item{dpthead}{String to become name of the column containing categories (\code{DPTVAL}) in output.}

\item{font}{Font face for text inside table}

\item{fontsize}{Font size for text inside table}

\item{boldheadyn}{Y/N to determine if table header should be bold}
}
\value{
flextable object
}
\description{
Create flextable output from display templates
}
\examples{
mentry_df <- adsl |>
  mentry(
    subset = NA_character_,
    byvar = NA_character_,
    trtvar = "TRT01A",
    trtsort = "TRT01AN",
    subgrpvar = "SEX/SEXN",
    trttotalyn = "N",
    add_grpmiss = "N",
    sgtotalyn = "N",
    pop_fil = "SAFFL"
  )
repdata <- adsl_summary(
  datain = mentry_df,
  vars = "AGEGR1~AGE-S~RACE"
) |>
  display_bign_head(mentry_df) |>
  tbl_processor(
    dptlabel = "Age Group~Age~Race",
    statlabel = NA,
    addrowvars = "DPTVAR"
  )
tbl_display(
  datain = repdata,
  bylabel = NA,
  dpthead = "  "
)
}
