% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_display.R
\name{tbl_processor}
\alias{tbl_processor}
\title{Prepare data for general table display}
\usage{
tbl_processor(
  datain,
  dptlabel = NA,
  statlabel = NA,
  extra_df = NULL,
  extra_mergeby = "DPTVAL",
  dropvars = "",
  disp_value_col = "Y",
  addrowvars = NA_character_
)
}
\arguments{
\item{datain}{Input dataframe}

\item{dptlabel}{Tilde-separated labels to set to category variables in data.
eg: If analysis vars (\code{DPTVAR}) contains \verb{AGEGR1, RACE}; \code{dptlabel} may be \code{"Age Group~Race"}.}

\item{statlabel}{Tilde-separated labels corresponding to each Statistic in data.}

\item{extra_df}{Additional dataframe, to merge with \code{datain}.}

\item{extra_mergeby}{Variables to merge \code{extra_df} by, if present.}

\item{dropvars}{Variables additional to standard present in input data to be removed from output}

\item{disp_value_col}{Hide/Show value column from the final display.}

\item{addrowvars}{Group Variable(s) to be removed as a column and instead used as row headers
in the value column.}
}
\value{
dataframe prepared for display using flextable
}
\description{
Prepare data for general table display
}
\examples{
mentry_df <- adsl |>
  mentry(
    subset = NA_character_,
    byvar = "ETHNIC/ETHNICN~BMIBLGR1",
    trtvar = "TRT01A",
    trtsort = "TRT01AN",
    subgrpvar = "SEX",
    trttotalyn = "N",
    add_grpmiss = "N",
    sgtotalyn = "N",
    pop_fil = "SAFFL"
  )
adsl_sum <- adsl_summary(
  datain = mentry_df,
  vars = "AGEGR1/AGEGR1N~AGE-S~RACE/RACEN"
) |>
  display_bign_head(mentry_df)
tbl_processor(
  datain = adsl_sum,
  dptlabel = "Age Group~Age~Race",
  statlabel = NA,
  addrowvars = "DPTVAR"
)
}
