% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{tbl_to_plot}
\alias{tbl_to_plot}
\title{Convert dataframe into ggplot object table}
\usage{
tbl_to_plot(
  datain,
  xvar = "XVAR",
  yvar,
  labelvar = yvar,
  colorvar = xvar,
  colors = rep("black", 10),
  text_size = 2.4,
  axis_opts = list(xsize = 8, xtsize = 6, xaxis_label = "")
)
}
\arguments{
\item{datain}{Input dataframe}

\item{xvar}{X axis variable, acts as table columns}

\item{yvar}{Y axis variable, acts as values within columns}

\item{labelvar}{Label variable}

\item{colorvar}{Variable to color values by}

\item{colors}{Vector of colors to be used per \code{colorvar}}

\item{text_size}{Size of text printed in 'table'}

\item{axis_opts}{Axis options, list containing \code{xsize}, and \code{xaxis_label} for X axis title;
Include \code{xtsize} element for ticks (column headers).}
}
\value{
ggplot object
}
\description{
Convert dataframe into ggplot object table
}
\examples{
MPG <- ggplot2::mpg
MPG[["cyl"]] <- as.character(MPG[["cyl"]])
tbl_to_plot(
  MPG,
  "cyl",
  "manufacturer",
  "manufacturer"
)
}
