% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{theme_cleany}
\alias{theme_cleany}
\title{Theme for ggplots with only X axis title/ticks}
\usage{
theme_cleany(
  xsize = 8,
  xtsize = 6,
  titlesize = 10,
  legend_opts = list(pos = "none", dir = "horizontal"),
  lines_color = "black"
)
}
\arguments{
\item{xsize}{Size of X axis label/title text}

\item{xtsize}{Size of X axis tick labels text}

\item{titlesize}{Size of plot title}

\item{legend_opts}{Legend styling option, a \code{list} containing \code{pos}(position) and
\code{dir} (direction).}

\item{lines_color}{Color to assign to graph panel border as well as axes lines}
}
\value{
A minimal ggplot2 theme without Y axis details
}
\description{
When this theme is applies to a plot, results in a white-background, full outer borders, no
Y axis label or breaks and center-aligned title.
}
\examples{
data(adsl)
library(ggplot2)
ggplot(
  adsl,
  aes(
    x = .data[["RACE"]], y = .data[["AGE"]],
    color = .data[["SEX"]]
  )
) +
  geom_point() +
  theme_cleany(legend_opts = list(pos = "bottom", dir = "horizontal"))

}
