% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{theme_std}
\alias{theme_std}
\title{Standard ggplot theme}
\usage{
theme_std(
  axis_opts = plot_axis_opts(),
  legend_opts = list(pos = "bottom", dir = "horizontal"),
  griddisplay = "N"
)
}
\arguments{
\item{axis_opts}{A \code{list} of axis specific options retrieved from \code{plot_axis_opts()}.}

\item{legend_opts}{Legend styling option, a \code{list} containing \code{pos}(position) and
\code{dir} (direction).}

\item{griddisplay}{Values "Y"/"N" to display grid lines}
}
\value{
Standard ggplot theme for basic figures using axis and legend options
}
\description{
Standard ggplot theme
}
\examples{
data(adsl)
library(ggplot2)
ggplot(
  adsl,
  aes(
    x = .data[["RACE"]], y = .data[["AGE"]],
    color = .data[["SEX"]]
  )
) +
  geom_point() +
  theme_std()
}
