% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tornado_plot.R
\name{tornado_plot}
\alias{tornado_plot}
\title{Tornado Plot}
\usage{
tornado_plot(
  datain,
  trt_left_label,
  trt_right_label,
  bar_width = 0.5,
  axis_opts = plot_axis_opts(xaxis_label = "Primary System Organ Class", yaxis_label =
    "\% of Subjects", ylinearopts = list(breaks = seq(-100, 100, 10), labels =
    (c(seq(100, 0, -10), seq(10, 100, 10))))),
  legend_opts = list(label = "Severity", pos = c(0.15, 0.15), dir = "vertical"),
  series_opts,
  griddisplay = "N"
)
}
\arguments{
\item{datain}{An input dataframe retrieved from \code{process_tornado_data}()`.}

\item{trt_left_label}{A Treatment label for displaying left hand side plot.}

\item{trt_right_label}{A Treatment label for displaying right hand side plot.}

\item{bar_width}{Width of individual bars \emph{numeric}}

\item{axis_opts}{A \code{list} of axis specific options retrieved from \code{plot_axis_opts()}.}

\item{legend_opts}{Legend styling option, a \code{list} containing \code{label}, \code{pos}(position) and
\code{dir} (direction).}

\item{series_opts}{Series Variable styling options, a \code{list} containing
\code{color} and \code{contrast}.}

\item{griddisplay}{Display Grid \code{(Y/N)}.}
}
\value{
Plot object
}
\description{
Tornado Plot
}
\examples{
data(tornado_plot_data)

tornado_df <- process_tornado_data(
  dataset_adsl = tornado_plot_data[["adsl"]],
  dataset_analysis = tornado_plot_data[["adae"]],
  adsl_subset = "SAFFL == 'Y'",
  analysis_subset = NA_character_,
  ae_filter = "Treatment emergent",
  obs_residual = "30",
  fmq_data = NA,
  ae_catvar = "AESEV",
  trtvar = "ARMCD",
  trt_left = "A",
  trt_right = "A",
  pop_fil = "Overall Population",
  pctdisp = "TRT",
  denom_subset = NA_character_,
  legendbign = "N",
  yvar = "AESOC"
)

series_opts <- g_seriescol(tornado_df, "blue~yellow~red", "BYVAR1")

tornado_plot(
  tornado_df,
  trt_left_label = "DRUG B",
  trt_right_label = "DRUG C",
  bar_width = 0.5,
  axis_opts = plot_axis_opts(
    xaxis_label = "Primary System Organ Class",
    yaxis_label = "\% of Subjects",
    ylinearopts = list(
      breaks = seq(-100, 100, 10),
      labels = c(seq(100, 0, -10), seq(10, 100, 10))
    )
  ),
  legend_opts = list(
    label = "Severity",
    pos = c(0.15, 0.15),
    dir = "vertical"
  ),
  series_opts = series_opts,
  griddisplay = "N"
)
}
