% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_pre_processor.R
\name{ae_pre_processor}
\alias{ae_pre_processor}
\title{Pre-processing data for Adverse Event domain specific reports}
\usage{
ae_pre_processor(
  datain,
  fmq_data = NULL,
  date_vars = c("ASTDT", "AENDT", "TRTSDT", "TRTEDT"),
  ae_filter = "Any Event",
  obs_residual = NA_real_
)
}
\arguments{
\item{datain}{the input ADaM (\emph{ADAE}) dataset.}

\item{fmq_data}{FMQ table dataframe, if passed, will be merged to adae date by PT.}

\item{date_vars}{Vector of variables to be converted to R date format}

\item{ae_filter}{Vector of adverse event types to be used as filter conditions.
Permissible Values: "ANY", "ANY EVENT", "TREATMENT EMERGENT", "SERIOUS",
"DRUG-RELATED", "RELATED", "MILD", "MODERATE", "SEVERE", "RECOVERED/RESOLVED",
"RECOVERING/RESOLVING", "NOT RECOVERING/NOT RESOLVING", "FATAL", "GRADE N"}

\item{obs_residual}{If not NA, use this argument to pass a period (numeric) to extend the
observation period. If passed as NA, overall study duration is considered for analysis.
eg. if 5, only events occurring upto 5 days past the TRTEDT are considered.}
}
\value{
: a list containing 2 objects
\itemize{
\item data - Processed dataframe output for further utilities (pass to \code{mentry()})
\item a_subset - Analysis Subset condition as a string
(pass to \code{mentry()} OR further utilities as required)
}
}
\description{
Pre-processing data for Adverse Event domain specific reports
}
\examples{
data(adae)
data(FMQ_Consolidated_List)

data_pre <- ae_pre_processor(
  datain = adae,
  ae_filter = "Serious",
  obs_residual = 0,
  fmq_data = FMQ_Consolidated_List
)
data_pre$data
data_pre$a_subset
}
