% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_volcano_plot.R
\name{ae_volcano_opts}
\alias{ae_volcano_opts}
\title{Volcano Plot axis Options}
\usage{
ae_volcano_opts(
  datain,
  trt1_label = "Control",
  trt2_label = "Exposure",
  statistic = "Risk Ratio",
  pvalue_trans = "none",
  xref_offset = 1
)
}
\arguments{
\item{datain}{Input data for volcano plot}

\item{trt1_label}{Control Treatment label}

\item{trt2_label}{Other treatment label}

\item{statistic}{Statistic used for risk analysis}

\item{pvalue_trans}{Required pvalue transformation. Values: "none"/"-log10"}

\item{xref_offset}{Offset value for X axis reference line; offset from 0 is \code{statistic} is Risk
Difference and from 1 if \code{statistic} is Risk Ratio.}
}
\value{
List of volcano-specific options
}
\description{
Volcano Plot axis Options
}
\examples{
data("adae")

ae_pre <- ae_pre_processor(
  datain = adae,
  obs_residual = 0,
  fmq_data = NA
)

ae_entry <- mentry(
  datain = ae_pre$data,
  subset = NA,
  byvar = "AEBODSYS",
  trtvar = "TRTA",
  trtsort = "TRTAN",
  subgrpvar = NA,
  trttotalyn = "N",
  add_grpmiss = "N",
  sgtotalyn = "N",
  pop_fil = "SAFFL"
)

ae_risk <- risk_stat(
  datain = ae_entry,
  a_subset = ae_pre$a_subset,
  summary_by = "Patients",
  eventvar = "AEDECOD",
  ctrlgrp = "Placebo",
  trtgrp = "Xanomeline High Dose",
  statistics = "Risk Ratio",
  alpha = 0.05,
  cutoff_where = "FREQ >5",
  sort_opt = "Ascending",
  sort_var = "Count"
)
ae_volcano_opts(ae_risk,
  pvalue_trans = "-log10"
)
}
