% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_merge.R
\name{dataset_merge}
\alias{dataset_merge}
\title{Merge Datasets}
\usage{
dataset_merge(..., byvars, subset = NULL)
}
\arguments{
\item{...}{Datasets to be merged.}

\item{byvars}{By variables required to perform merge.}

\item{subset}{Dataset specific subset conditions as \code{list}, default is \code{NULL}.
Has to be specified in the same order of datasets to be merged}
}
\value{
A \code{data.frame}
}
\description{
Merge Datasets
}
\examples{
dataset_merge(
  lab_data$adsl,
  lab_data$adlb,
  byvars = "STUDYID~USUBJID~SUBJID",
  subset = list("SEX=='F'", "PARAMCD == 'L00021S'")
)

dataset_merge(
  lab_data$adsl,
  lab_data$adlb,
  byvars = "STUDYID~USUBJID~SUBJID",
  subset = list("SEX=='F'", NA_character_)
)

dataset_merge(
  lab_data$adsl,
  lab_data$adlb,
  byvars = "STUDYID~USUBJID~SUBJID",
  subset = list(NA_character_, "PARAMCD == 'L00021S'")
)

dataset_merge(
  lab_data$adsl,
  lab_data$adlb,
  byvars = "STUDYID~USUBJID~SUBJID",
  subset = list("USUBJID == 'XYZ1 1003 10031009'", NA_character_)
)

dataset_merge(
  waterfall_plot_data$adrs,
  waterfall_plot_data$adtr,
  byvars = "STUDYID~USUBJID~TRT01P",
  subset = list("PARAMCD == 'BOR_C'", NA_character_)
)

## more than 2 datasets

dataset_merge(
  dplyr::filter(lab_data$adsl, USUBJID == "XYZ1 1003 10031009"),
  lab_data$adsl,
  lab_data$adlb,
  byvars = "STUDYID~USUBJID~SUBJID"
)

}
