% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adlb_r301.R, R/edish_plot.R
\name{lab_abnormality_summary}
\alias{lab_abnormality_summary}
\title{Incidence of Laboratory Test Abnormalities (Without Regard to Baseline Abnormality)}
\usage{
lab_abnormality_summary(
  datain,
  crit_vars = "CRIT3~CRIT4",
  pctdisp = "SUBGRP",
  a_subset = NA_character_,
  denom_subset = NA_character_
)

lab_abnormality_summary(
  datain,
  crit_vars = "CRIT3~CRIT4",
  pctdisp = "SUBGRP",
  a_subset = NA_character_,
  denom_subset = NA_character_
)
}
\arguments{
\item{datain}{Input dataset (\code{adlb}).}

\item{crit_vars}{Criteria variables}

\item{pctdisp}{Denominator to calculate percentages by.
Values: \verb{"TRT", "VAR","COL", "SUBGRP", "SGRPN", "CAT", "NONE", "NO", "DPTVAR"}}

\item{a_subset}{Subset conditions for analysis of dependent variable.}

\item{denom_subset}{Subset conditions for denominator eg. \code{"APSBLFL == 'Y'"}}
}
\value{
\code{data.frame} with summary of laboratory abnormality incidence counts

\code{data.frame} with summary of laboratory abnormality incidence counts
}
\description{
Incidence of Laboratory Test Abnormalities (Without Regard to Baseline Abnormality)

Incidence of Laboratory Test Abnormalities (Without Regard to Baseline Abnormality)
}
\examples{
data("lab_data")

lb_entry <- lab_data$adlb |>
  mentry(
    subset = NA_character_,
    byvar = "PARCAT1~PARAM",
    subgrpvar = NA_character_,
    trtvar = "TRTA",
    trtsort = "TRTAN",
    trttotalyn = "N",
    sgtotalyn = "N",
    add_grpmiss = "N",
    pop_fil = "SAFFL"
  )

out <-
  lb_entry |>
  lab_abnormality_summary(
    crit_vars = "CRIT3~CRIT4",
    pctdisp = "SUBGRP",
    a_subset = NA_character_,
    denom_subset = NA_character_
  ) |>
  display_bign_head(mentry_data = lb_entry) |>
  tbl_processor(
    dptlabel = ""
  )

out

# `flextable` output
out |>
  tbl_display(
    bylabel = "Parameter Category~Parameter",
    dpthead = "Primary Criteria"
  )

data("lab_data")

lb_entry <- lab_data$adlb |>
  mentry(
    subset = NA_character_,
    byvar = "PARCAT1~PARAM",
    subgrpvar = NA_character_,
    trtvar = "TRTA",
    trtsort = "TRTAN",
    trttotalyn = "N",
    sgtotalyn = "N",
    add_grpmiss = "N",
    pop_fil = "SAFFL"
  )

out <-
  lb_entry |>
  lab_abnormality_summary(
    crit_vars = "CRIT3~CRIT4",
    pctdisp = "SUBGRP",
    a_subset = NA_character_,
    denom_subset = NA_character_
  ) |>
  display_bign_head(mentry_data = lb_entry) |>
  tbl_processor(
    dptlabel = ""
  )

out

# `flextable` output
out |>
  tbl_display(
    bylabel = "Parameter Category~Parameter",
    dpthead = "Primary Criteria"
  )

}
