% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_plot.R
\name{line_plot}
\alias{line_plot}
\title{Utility for Line Plot}
\usage{
line_plot(
  datain,
  series_var = "TRTVAR",
  series_labelvar = series_var,
  series_opts,
  axis_opts = plot_axis_opts(),
  legend_opts = list(label = "", pos = "bottom", dir = "horizontal"),
  griddisplay = "N",
  plot_title = NULL,
  dodge_width = NULL
)
}
\arguments{
\item{datain}{Input dataset from process_line_plot_data() output.}

\item{series_var}{Series (Treatment) Variable.}

\item{series_labelvar}{Series (Treatment) Variable labels for legend, if
different from \code{series_var}. If this is needed, then \code{series_labelvar}
should be a factor variable with levels corresponding to \code{series_var}, also a factor.}

\item{series_opts}{Series Variable styling options, a \code{list} containing
\code{shape}, \code{color} and \code{size}.}

\item{axis_opts}{A \code{list} of axis specific options retrieved from \code{plot_axis_opts()}.}

\item{legend_opts}{Legend styling option, a \code{list} containing \code{label}, \code{pos}(position) and
\code{dir} (direction).}

\item{griddisplay}{Display Grid \code{(Y/N)}.}

\item{plot_title}{Text to use as plot title, if required}

\item{dodge_width}{Width to dodge points/lines by, IF required.}
}
\value{
plot - Line plot.
}
\description{
Utility for Line Plot
}
\examples{
data("vx_line_data")

lineplot_df <- process_line_plot_data(
  dataset_adsl = vx_line_data$adsl,
  dataset_analysis = vx_line_data$adva,
  adsl_subset = "SAFFL == 'Y'",
  analysis_subset = 'PARAMN==2 & TRTARN!=""& (AVISITN \%in\% c(1,3,4,5,6,7))',
  trtvar = "TRTP",
  trtsort = "TRTPN",
  xvar = "AVISIT",
  yvar = "AVAL"
)

plot <- line_plot(
  datain = lineplot_df,
  series_var = "TRTVAR",
  series_labelvar = "TRTVAR",
  series_opts = list(
    color = g_seriescol(lineplot_df, NA, "TRTVAR")
  ),
  axis_opts = plot_axis_opts(
    ylinearopts = list(
      breaks = c(100, 1000, 10000, 100000),
      limits = c(100, 100000)
    ),
    xaxis_label = "Visit",
    yaxis_label = "Geometric Mean Titer"
  ),
  legend_opts = list(
    label = "",
    pos = "bottom",
    dir = "horizontal"
  ),
  griddisplay = "N",
  plot_title = "Line Plot"
)

plot

}
