% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vx_boxplot.R
\name{process_vx_box_data}
\alias{process_vx_box_data}
\title{Process data for vaccine boxplot}
\usage{
process_vx_box_data(
  dataset_adsl,
  dataset_analysis,
  adsl_subset = "SAFFL == 'Y'",
  analysis_subset = "((ANL08FL == 'Y')|(ANL09FL=='Y'))&(ARMCD!='')",
  split_by = NA_character_,
  trtvar = "TRT01A",
  trtsort = "TRT01AN",
  xvar = "AVISIT",
  yvar = "AVAL",
  ystat = "mean",
  legendbign = "Y",
  ada_nab_opts = list(N = "1~2", LAB = "ADA Titer (log2)~NAb Titer (log2)", REF =
    "6.23~1.58", DIL = "75~3")
)
}
\arguments{
\item{ystat}{Additional Statistic to be calculated and plotted as markers.
Values: 'mean', 'sum', 'sd' etc}

\item{ada_nab_opts}{List of values of : \emph{PARAMN}, Y axis Label, Reference
line value and Dilution (for footnote) corresponding to ADA and NAb titers
respectively. Format: list(N = "1~2",
LAB = "ADA Titer (log2)~NAb Titer (log2)", REF = "6.23~1.58", DIL = "75~3")}
}
\value{
Dataframe containing analysis values for requisite box plot statistics
}
\description{
Process data for vaccine boxplot
}
\examples{
data(vx_box_data)
process_vx_box_data(
  dataset_adsl = vx_box_data$adsl,
  dataset_analysis = vx_box_data$adisda,
  adsl_subset = "RANDFL == 'Y'",
  analysis_subset = "((ANL08FL == 'Y')|(ANL09FL=='Y'))&(ARMCD!='')&(PARAMN \%in\% c(1, 2))",
  trtvar = "TRTA",
  trtsort = "TRTAN",
  xvar = "AVISIT",
  ystat = "mean"
)
}
