% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schoenfeld_plot.R
\name{schoenfeld_plot}
\alias{schoenfeld_plot}
\title{Schoenfeld Plot}
\usage{
schoenfeld_plot(
  datain,
  disp_clmband = "Y",
  ties_method = "efron",
  time_unit = "Months",
  markersize = 3,
  marker_symbol = shape_to_sym("circle"),
  axis_opts = plot_axis_opts(),
  df = 2,
  pvalue_decimal = 4,
  pair_id = NA_character_
)
}
\arguments{
\item{datain}{input dataset derived from \code{surv_pre_processor()}}

\item{disp_clmband}{Display CLM band in graph (\code{Y/N}) Default: \code{Y}}

\item{ties_method}{Method of handling ties in ph reg model statement.
Default: \code{"efron"}. Values: \verb{"efron", "breslow", "exact"} (\code{"exact"} is not
appropriate when the time variable is continuous or in presence of tied
events, see more here \link[survival]{coxph}).}

\item{time_unit}{Time unit to be displayed in the X axis label.
Default: "YEARS"}

\item{markersize}{Size of the marker to be displayed in the plot.
Default:3}

\item{marker_symbol}{Symbol for residual points from the utility
\code{shape_to_sym()}.Default: \code{"circle"}. Values: \code{"circle"/"circlefilled"}}

\item{axis_opts}{A \code{list} of axis specific options retrieved from
\code{plot_axis_opts()}}

\item{df}{Degrees of Freedom for the fitted natural spline.}

\item{pvalue_decimal}{Number of decimals needed for P-value.
Default: 4}

\item{pair_id}{Select the number of pairs required.User can give input as
only one set of pairs.For multiple pairs user have to do other subsetting.}
}
\value{
List containing Schoenfeld plots and Cox Regression statistics
\itemize{
\item pval: p-value
\item pair: Treatment pairs
\item s_plot: Schoenfeld Plot object
}
}
\description{
Schoenfeld Plot
}
\examples{


data("survival")

sh_pre <- surv_pre_processor(
  dataset_adsl = survival$adsl,
  adsl_subset = "SAFFL=='Y'",
  dataset_analysis = survival$adtte,
  split_by = NA_character_,
  analysis_subset = "PARAMCD=='OS' & FASFL=='Y'",
  trtsort = "TRT01PN",
  censor_var = "CNSR",
  censor_val = 1,
  trtvar = "TRT01P",
  time_var = "AVAL"
)

sh_plot <- schoenfeld_plot(
  datain = sh_pre,
  disp_clmband = "Y",
  ties_method = "efron",
  time_unit = "Months",
  markersize = 1.5,
  marker_symbol = shape_to_sym("circle"),
  axis_opts = plot_axis_opts(),
  pvalue_decimal = 4,
  pair_id = "1~2"
)
sh_plot

}
