% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary_functions}
\alias{summary_functions}
\title{Create summary stats function for use within \code{msumstat()}}
\usage{
summary_functions(sigdec = 2)
}
\arguments{
\item{sigdec}{Number of significant decimal places (base)}
}
\value{
a named list containing function definition for all defined summary
statistics - mean, min, max, median, iqr, var, sum, sd, q25, q75, p1, p5,
p10, p90, p95, p99 (where last digits represent \% of quantile), meansd,
range, q1q3, medianrange (concatenation of indicated names), whiskerlow,
whiskerup, outliers in the Tukey method for box statistics
}
\description{
Create summary stats function for use within \code{msumstat()}
}
